Updating Javascript build
