+++ view/templates/_master.twig
@@ -15,13 +15,10 @@
     <link rel="shortcut icon" href="{{ asset('images/favicon-bolt.png') }}">
     <link href='//fonts.googleapis.com/css?family=Source+Sans+Pro:300,400,600,600italic,400italic' rel='stylesheet' type='text/css'>
 
-    <!-- Highlighter -->
-    <script src="{{ asset('js/highlight.pack.js') }}"></script>
-
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/docsearch.js/1/docsearch.min.css" />
     <link rel="stylesheet" href="{{ asset('styles/docs.css') }}" type="text/css" media="all" />
-    <script src="{{ asset('js/modernizr.js') }}"></script>
     <script src="{{ asset('js/jquery.js') }}"></script>
+    <script src="{{ asset('js/docs.js') }}"></script>
 
 {% block extrahead %}
 {% endblock extrahead %}
@@ -105,11 +102,6 @@
 {% endblock content %}
 
     </main>
-
-    <script src="{{ asset('js/jquery.magnific-popup.min.js') }}"></script>
-    <script src="{{ asset('js/clipboard.min.js') }}"></script>
-    <script src="{{ asset('js/js.cookie.js') }}"></script>
-    <script src="{{ asset('js/tree.jquery.js') }}"></script>
     <script>var prefix = "";</script>
     <script src="{{ asset('js/docs.js') }}" type="text/javascript"></script>
     <script type="text/javascript" src="https://cdn.jsdelivr.net/docsearch.js/1/docsearch.min.js"></script>
