+++ web/js/docs.js
@@ -1,152 +1,4 @@
-
-var q = "";
-
-jQuery(function($) {
-
-    hljs.initHighlightingOnLoad();
-
-    // hiding the mobile navigation
-    $('.main-nav').removeClass('expanded');
-
-    // and toggling it again with a button
-    $('.menu-toggle').click(function() {
-        $('.main-nav').toggleClass('expanded');
-        $(this).toggleClass('active');
-    });
-
-    $('main .content a.popup').magnificPopup({
-        type: 'image'
-    });
-
-    $('div.gallery-popup').magnificPopup({
-      delegate: 'a',
-      type: 'image',
-      gallery:{
-        enabled:true
-      }
-    });
-
-    $(window).scroll(function () {
-        $('header').css('backgroundPosition', '0px ' + (posTop() / 2) + 'px');
-    });
-
-    var pathname = window.location.pathname;
-
-    var $tree = $('#tree1');
-    $tree.tree({
-        saveState: 'boltmenu',
-        data: $tree.data('data'),
-        onCreateLi: function(node, $li) {
-            $li.toggleClass('jqtree-selected', node.url === pathname);
-
-            if (pathname !== node.url && pathname.indexOf(node.url) === 0) {
-                $tree.tree('openNode', node);
-                $li.removeClass('jqtree-closed');
-            }
-        }
-    });
-
-    $('#tree1').bind(
-        'tree.click',
-        function(event) {
-            // The clicked node is 'event.node'
-            var node = event.node;
-            var theURL = node.url;
-            if (theURL) {
-                location.href = theURL;
-            }
-            $tree.tree('toggle', node);
-        }
-    );
-
-    $('#jqtree-collapse-all').click(function() {
-        var tree = $tree.tree('getTree');
-
-        tree.iterate(function(node, level) {
-            if (node.hasChildren()) {
-                // This will close the folder
-                $tree.tree('closeNode', node);
-                return false;
-            }
-
-            return true;
-        });
-
-    });
-
-    $('#jqtree-expand-all').click(function() {
-        var tree = $tree.tree('getTree');
-
-        tree.iterate(function(node, level) {
-            if (node.hasChildren()) {
-                // This will open the folder
-                $tree.tree('openNode', node);
-                return false;
-            }
-
-            return true;
-        });
-
-    });
-
-    // Update the number of stars. Stolen from foundation.zurb.com.
-    $.ajax({
-      dataType: 'jsonp',
-      url: 'https://api.github.com/repos/bolt/bolt?callback=boltGithub&access_token=8e0dfc559d22265208b2924266c8b15b60fd9b85',
-      success: function (response) {
-        if (response && response.data.watchers) {
-          var watchers = response.data.watchers;
-          // var watchers = (Math.round((response.data.watchers / 100), 10) / 10).toFixed(1);
-          $('#stargazers').html(watchers + ' Stars');
-        }
-      }
-    });
-
-    // Jumpmenu for the versions
-    $("#version-changer-submit").hide();
-    $("#version-changer select").change(function() {
-        console.log('jo');
-        window.location = $("#version-changer select option:selected").val();
-    })
-
-    //Zero Clipboard stuff..
-    $('pre code').each(function(index) {
-        // Get the text to be copied to the clipboard
-        var text = $(this).text();
-
-        // Create the copy button
-        var copyBtn = $('<span class="copy-btn">[ Copy Code ]</span>')
-            .attr('data-clipboard-text', text) // set the text to be copied
-            .insertBefore(this); // insert copy button before <pre>
-    });
-
-    initClipBoard();
-
-});
-
-function initClipBoard() {
-
-    var clipboard = new Clipboard('.copy-btn');
-
-    clipboard.on('success', function(e) {
-        $(e.trigger).text('[ Copied ]');
-        window.setTimeout(function(){ $(e.trigger).text('[ Copy code ]'); }, 2000);
-        e.clearSelection();
-    });
-
-}
-
-function formatForUrl(str) {
-    return str.replace(/_/g, '-')
-        .replace(/ /g, '-')
-        .replace(/:/g, '-')
-        .replace(/\\/g, '-')
-        .replace(/\//g, '-')
-        .replace(/[^a-zA-Z0-9\-]+/g, '')
-        .replace(/-{2,}/g, '-')
-        .toLowerCase();
-};
-
-function posTop() {
-    return typeof window.pageYOffset != 'undefined' ? window.pageYOffset: document.documentElement.scrollTop? document.documentElement.scrollTop: document.body.scrollTop? document.body.scrollTop:0;
-}
+function initClipBoard(){var e=new Clipboard(".copy-btn");e.on("success",function(e){$(e.trigger).text("[ Copied ]"),window.setTimeout(function(){$(e.trigger).text("[ Copy code ]")},2e3),e.clearSelection()})}function formatForUrl(e){return e.replace(/_/g,"-").replace(/ /g,"-").replace(/:/g,"-").replace(/\\/g,"-").replace(/\//g,"-").replace(/[^a-zA-Z0-9\-]+/g,"").replace(/-{2,}/g,"-").toLowerCase()}function posTop(){return"undefined"!=typeof window.pageYOffset?window.pageYOffset:document.documentElement.scrollTop?document.documentElement.scrollTop:document.body.scrollTop?document.body.scrollTop:0}!function(e){if("object"==typeof exports&&"undefined"!=typeof module)module.exports=e();else if("function"==typeof define&&define.amd)define([],e);else{var t;t="undefined"!=typeof window?window:"undefined"!=typeof global?global:"undefined"!=typeof self?self:this,t.Clipboard=e()}}(function(){return function e(t,n,r){function o(s,a){if(!n[s]){if(!t[s]){var l="function"==typeof require&&require;if(!a&&l)return l(s,!0);if(i)return i(s,!0);var c=new Error("Cannot find module '"+s+"'");throw c.code="MODULE_NOT_FOUND",c}var u=n[s]={exports:{}};t[s][0].call(u.exports,function(e){var n=t[s][1][e];return o(n?n:e)},u,u.exports,e,t,n,r)}return n[s].exports}for(var i="function"==typeof require&&require,s=0;s<r.length;s++)o(r[s]);return o}({1:[function(e,t,n){var r=e("matches-selector");t.exports=function(e,t,n){for(var o=n?e:e.parentNode;o&&o!==document;){if(r(o,t))return o;o=o.parentNode}}},{"matches-selector":2}],2:[function(e,t,n){function r(e,t){if(i)return i.call(e,t);for(var n=e.parentNode.querySelectorAll(t),r=0;r<n.length;++r)if(n[r]==e)return!0;return!1}var o=Element.prototype,i=o.matchesSelector||o.webkitMatchesSelector||o.mozMatchesSelector||o.msMatchesSelector||o.oMatchesSelector;t.exports=r},{}],3:[function(e,t,n){function r(e,t,n,r){var i=o.apply(this,arguments);return e.addEventListener(n,i),{destroy:function(){e.removeEventListener(n,i)}}}function o(e,t,n,r){return function(n){var o=i(n.target,t,!0);o&&(Object.defineProperty(n,"target",{value:o}),r.call(e,n))}}var i=e("closest");t.exports=r},{closest:1}],4:[function(e,t,n){n.node=function(e){return void 0!==e&&e instanceof HTMLElement&&1===e.nodeType},n.nodeList=function(e){var t=Object.prototype.toString.call(e);return void 0!==e&&("[object NodeList]"===t||"[object HTMLCollection]"===t)&&"length"in e&&(0===e.length||n.node(e[0]))},n.string=function(e){return"string"==typeof e||e instanceof String},n["function"]=function(e){var t=Object.prototype.toString.call(e);return"[object Function]"===t}},{}],5:[function(e,t,n){function r(e,t,n){if(!e&&!t&&!n)throw new Error("Missing required arguments");if(!a.string(t))throw new TypeError("Second argument must be a String");if(!a["function"](n))throw new TypeError("Third argument must be a Function");if(a.node(e))return o(e,t,n);if(a.nodeList(e))return i(e,t,n);if(a.string(e))return s(e,t,n);throw new TypeError("First argument must be a String, HTMLElement, HTMLCollection, or NodeList")}function o(e,t,n){return e.addEventListener(t,n),{destroy:function(){e.removeEventListener(t,n)}}}function i(e,t,n){return Array.prototype.forEach.call(e,function(e){e.addEventListener(t,n)}),{destroy:function(){Array.prototype.forEach.call(e,function(e){e.removeEventListener(t,n)})}}}function s(e,t,n){return l(document.body,e,t,n)}var a=e("./is"),l=e("delegate");t.exports=r},{"./is":4,delegate:3}],6:[function(e,t,n){function r(e){var t;if("INPUT"===e.nodeName||"TEXTAREA"===e.nodeName)e.select(),t=e.value;else{var n=window.getSelection(),r=document.createRange();r.selectNodeContents(e),n.removeAllRanges(),n.addRange(r),t=n.toString()}return t}t.exports=r},{}],7:[function(e,t,n){function r(){}r.prototype={on:function(e,t,n){var r=this.e||(this.e={});return(r[e]||(r[e]=[])).push({fn:t,ctx:n}),this},once:function(e,t,n){function r(){o.off(e,r),t.apply(n,arguments)}var o=this;return r._=t,this.on(e,r,n)},emit:function(e){var t=[].slice.call(arguments,1),n=((this.e||(this.e={}))[e]||[]).slice(),r=0,o=n.length;for(r;o>r;r++)n[r].fn.apply(n[r].ctx,t);return this},off:function(e,t){var n=this.e||(this.e={}),r=n[e],o=[];if(r&&t)for(var i=0,s=r.length;s>i;i++)r[i].fn!==t&&r[i].fn._!==t&&o.push(r[i]);return o.length?n[e]=o:delete n[e],this}},t.exports=r},{}],8:[function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{"default":e}}function o(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}n.__esModule=!0;var i=function(){function e(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}return function(t,n,r){return n&&e(t.prototype,n),r&&e(t,r),t}}(),s=e("select"),a=r(s),l=function(){function e(t){o(this,e),this.resolveOptions(t),this.initSelection()}return e.prototype.resolveOptions=function(){var e=arguments.length<=0||void 0===arguments[0]?{}:arguments[0];this.action=e.action,this.emitter=e.emitter,this.target=e.target,this.text=e.text,this.trigger=e.trigger,this.selectedText=""},e.prototype.initSelection=function(){if(this.text&&this.target)throw new Error('Multiple attributes declared, use either "target" or "text"');if(this.text)this.selectFake();else{if(!this.target)throw new Error('Missing required attributes, use either "target" or "text"');this.selectTarget()}},e.prototype.selectFake=function(){var e=this;this.removeFake(),this.fakeHandler=document.body.addEventListener("click",function(){return e.removeFake()}),this.fakeElem=document.createElement("textarea"),this.fakeElem.style.position="absolute",this.fakeElem.style.left="-9999px",this.fakeElem.style.top=(window.pageYOffset||document.documentElement.scrollTop)+"px",this.fakeElem.setAttribute("readonly",""),this.fakeElem.value=this.text,document.body.appendChild(this.fakeElem),this.selectedText=a["default"](this.fakeElem),this.copyText()},e.prototype.removeFake=function(){this.fakeHandler&&(document.body.removeEventListener("click"),this.fakeHandler=null),this.fakeElem&&(document.body.removeChild(this.fakeElem),this.fakeElem=null)},e.prototype.selectTarget=function(){this.selectedText=a["default"](this.target),this.copyText()},e.prototype.copyText=function(){var e=void 0;try{e=document.execCommand(this.action)}catch(t){e=!1}this.handleResult(e)},e.prototype.handleResult=function(e){e?this.emitter.emit("success",{action:this.action,text:this.selectedText,trigger:this.trigger,clearSelection:this.clearSelection.bind(this)}):this.emitter.emit("error",{action:this.action,trigger:this.trigger,clearSelection:this.clearSelection.bind(this)})},e.prototype.clearSelection=function(){this.target&&this.target.blur(),window.getSelection().removeAllRanges()},e.prototype.destroy=function(){this.removeFake()},i(e,[{key:"action",set:function(){var e=arguments.length<=0||void 0===arguments[0]?"copy":arguments[0];if(this._action=e,"copy"!==this._action&&"cut"!==this._action)throw new Error('Invalid "action" value, use either "copy" or "cut"')},get:function(){return this._action}},{key:"target",set:function(e){if(void 0!==e){if(!e||"object"!=typeof e||1!==e.nodeType)throw new Error('Invalid "target" value, use a valid Element');this._target=e}},get:function(){return this._target}}]),e}();n["default"]=l,t.exports=n["default"]},{select:6}],9:[function(e,t,n){"use strict";function r(e){return e&&e.__esModule?e:{"default":e}}function o(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}function i(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function, not "+typeof t);e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,enumerable:!1,writable:!0,configurable:!0}}),t&&(Object.setPrototypeOf?Object.setPrototypeOf(e,t):e.__proto__=t)}function s(e,t){var n="data-clipboard-"+e;return t.hasAttribute(n)?t.getAttribute(n):void 0}n.__esModule=!0;var a=e("./clipboard-action"),l=r(a),c=e("tiny-emitter"),u=r(c),d=e("good-listener"),p=r(d),h=function(e){function t(n,r){o(this,t),e.call(this),this.resolveOptions(r),this.listenClick(n)}return i(t,e),t.prototype.resolveOptions=function(){var e=arguments.length<=0||void 0===arguments[0]?{}:arguments[0];this.action="function"==typeof e.action?e.action:this.defaultAction,this.target="function"==typeof e.target?e.target:this.defaultTarget,this.text="function"==typeof e.text?e.text:this.defaultText},t.prototype.listenClick=function(e){var t=this;this.listener=p["default"](e,"click",function(e){return t.onClick(e)})},t.prototype.onClick=function(e){this.clipboardAction&&(this.clipboardAction=null),this.clipboardAction=new l["default"]({action:this.action(e.target),target:this.target(e.target),text:this.text(e.target),trigger:e.target,emitter:this})},t.prototype.defaultAction=function(e){return s("action",e)},t.prototype.defaultTarget=function(e){var t=s("target",e);return t?document.querySelector(t):void 0},t.prototype.defaultText=function(e){return s("text",e)},t.prototype.destroy=function(){this.listener.destroy(),this.clipboardAction&&(this.clipboardAction.destroy(),this.clipboardAction=null)},t}(u["default"]);n["default"]=h,t.exports=n["default"]},{"./clipboard-action":8,"good-listener":5,"tiny-emitter":7}]},{},[9])(9)});var q="";jQuery(function(e){hljs.initHighlightingOnLoad(),e(".main-nav").removeClass("expanded"),e(".menu-toggle").click(function(){e(".main-nav").toggleClass("expanded"),e(this).toggleClass("active")}),e("main .content a.popup").magnificPopup({type:"image"}),e("div.gallery-popup").magnificPopup({delegate:"a",type:"image",gallery:{enabled:!0}}),e(window).scroll(function(){e("header").css("backgroundPosition","0px "+posTop()/2+"px")});var t=window.location.pathname,n=e("#tree1");n.tree({saveState:"boltmenu",data:n.data("data"),onCreateLi:function(e,r){r.toggleClass("jqtree-selected",e.url===t),t!==e.url&&0===t.indexOf(e.url)&&(n.tree("openNode",e),r.removeClass("jqtree-closed"))}}),e("#tree1").bind("tree.click",function(e){var t=e.node,r=t.url;r&&(location.href=r),n.tree("toggle",t)}),e("#jqtree-collapse-all").click(function(){var e=n.tree("getTree");e.iterate(function(e,t){return e.hasChildren()?(n.tree("closeNode",e),!1):!0})}),e("#jqtree-expand-all").click(function(){var e=n.tree("getTree");e.iterate(function(e,t){return e.hasChildren()?(n.tree("openNode",e),!1):!0})}),e.ajax({dataType:"jsonp",url:"https://api.github.com/repos/bolt/bolt?callback=boltGithub&access_token=8e0dfc559d22265208b2924266c8b15b60fd9b85",success:function(t){if(t&&t.data.watchers){var n=t.data.watchers;e("#stargazers").html(n+" Stars")}}}),e("#version-changer-submit").hide(),e("#version-changer select").change(function(){console.log("jo"),window.location=e("#version-changer select option:selected").val()}),e("pre code").each(function(t){var n=e(this).text();e('<span class="copy-btn">[ Copy Code ]</span>').attr("data-clipboard-text",n).insertBefore(this)}),initClipBoard()}),!function(e){"undefined"!=typeof exports?e(exports):(window.hljs=e({}),"function"==typeof define&&define.amd&&define("hljs",[],function(){return window.hljs}))}(function(e){function t(e){return e.replace(/&/gm,"&amp;").replace(/</gm,"&lt;").replace(/>/gm,"&gt;")}function n(e){return e.nodeName.toLowerCase()}function r(e,t){var n=e&&e.exec(t);return n&&0==n.index}function o(e){return/^(no-?highlight|plain|text)$/i.test(e)}function i(e){var t,n,r,i=e.className+" ";if(i+=e.parentNode?e.parentNode.className:"",n=/\blang(?:uage)?-([\w-]+)\b/i.exec(i))return b(n[1])?n[1]:"no-highlight";for(i=i.split(/\s+/),t=0,r=i.length;r>t;t++)if(b(i[t])||o(i[t]))return i[t]}function s(e,t){var n,r={};for(n in e)r[n]=e[n];if(t)for(n in t)r[n]=t[n];return r}function a(e){var t=[];return function r(e,o){for(var i=e.firstChild;i;i=i.nextSibling)3==i.nodeType?o+=i.nodeValue.length:1==i.nodeType&&(t.push({event:"start",offset:o,node:i}),o=r(i,o),n(i).match(/br|hr|img|input/)||t.push({event:"stop",offset:o,node:i}));return o}(e,0),t}function l(e,r,o){function i(){return e.length&&r.length?e[0].offset!=r[0].offset?e[0].offset<r[0].offset?e:r:"start"==r[0].event?e:r:e.length?e:r}function s(e){function r(e){return" "+e.nodeName+'="'+t(e.value)+'"'}u+="<"+n(e)+Array.prototype.map.call(e.attributes,r).join("")+">"}function a(e){u+="</"+n(e)+">"}function l(e){("start"==e.event?s:a)(e.node)}for(var c=0,u="",d=[];e.length||r.length;){var p=i();if(u+=t(o.substr(c,p[0].offset-c)),c=p[0].offset,p==e){d.reverse().forEach(a);do l(p.splice(0,1)[0]),p=i();while(p==e&&p.length&&p[0].offset==c);d.reverse().forEach(s)}else"start"==p[0].event?d.push(p[0].node):d.pop(),l(p.splice(0,1)[0])}return u+t(o.substr(c))}function c(e){function t(e){return e&&e.source||e}function n(n,r){return new RegExp(t(n),"m"+(e.cI?"i":"")+(r?"g":""))}function r(o,i){if(!o.compiled){if(o.compiled=!0,o.k=o.k||o.bK,o.k){var a={},l=function(t,n){e.cI&&(n=n.toLowerCase()),n.split(" ").forEach(function(e){var n=e.split("|");a[n[0]]=[t,n[1]?Number(n[1]):1]})};"string"==typeof o.k?l("keyword",o.k):Object.keys(o.k).forEach(function(e){l(e,o.k[e])}),o.k=a}o.lR=n(o.l||/\b\w+\b/,!0),i&&(o.bK&&(o.b="\\b("+o.bK.split(" ").join("|")+")\\b"),o.b||(o.b=/\B|\b/),o.bR=n(o.b),o.e||o.eW||(o.e=/\B|\b/),o.e&&(o.eR=n(o.e)),o.tE=t(o.e)||"",o.eW&&i.tE&&(o.tE+=(o.e?"|":"")+i.tE)),o.i&&(o.iR=n(o.i)),void 0===o.r&&(o.r=1),o.c||(o.c=[]);var c=[];o.c.forEach(function(e){e.v?e.v.forEach(function(t){c.push(s(e,t))}):c.push("self"==e?o:e)}),o.c=c,o.c.forEach(function(e){r(e,o)}),o.starts&&r(o.starts,i);var u=o.c.map(function(e){return e.bK?"\\.?("+e.b+")\\.?":e.b}).concat([o.tE,o.i]).map(t).filter(Boolean);o.t=u.length?n(u.join("|"),!0):{exec:function(){return null}}}}r(e)}function u(e,n,o,i){function s(e,t){for(var n=0;n<t.c.length;n++)if(r(t.c[n].bR,e))return t.c[n]}function a(e,t){if(r(e.eR,t)){for(;e.endsParent&&e.parent;)e=e.parent;return e}return e.eW?a(e.parent,t):void 0}function l(e,t){return!o&&r(t.iR,e)}function p(e,t){var n=y.cI?t[0].toLowerCase():t[0];return e.k.hasOwnProperty(n)&&e.k[n]}function h(e,t,n,r){var o=r?"":w.classPrefix,i='<span class="'+o,s=n?"":"</span>";return i+=e+'">',i+t+s}function f(){if(!C.k)return t(E);var e="",n=0;C.lR.lastIndex=0;for(var r=C.lR.exec(E);r;){e+=t(E.substr(n,r.index-n));var o=p(C,r);o?(I+=o[1],e+=h(o[0],t(r[0]))):e+=t(r[0]),n=C.lR.lastIndex,r=C.lR.exec(E)}return e+t(E.substr(n))}function m(){var e="string"==typeof C.sL;if(e&&!N[C.sL])return t(E);var n=e?u(C.sL,E,!0,S[C.sL]):d(E,C.sL.length?C.sL:void 0);return C.r>0&&(I+=n.r),e&&(S[C.sL]=n.top),h(n.language,n.value,!1,!0)}function g(){return void 0!==C.sL?m():f()}function _(e,n){var r=e.cN?h(e.cN,"",!0):"";e.rB?(k+=r,E=""):e.eB?(k+=t(n)+r,E=""):(k+=r,E=n),C=Object.create(e,{parent:{value:C}})}function v(e,n){if(E+=e,void 0===n)return k+=g(),0;var r=s(n,C);if(r)return k+=g(),_(r,n),r.rB?0:n.length;var o=a(C,n);if(o){var i=C;i.rE||i.eE||(E+=n),k+=g();do C.cN&&(k+="</span>"),I+=C.r,C=C.parent;while(C!=o.parent);return i.eE&&(k+=t(n)),E="",o.starts&&_(o.starts,""),i.rE?0:n.length}if(l(n,C))throw new Error('Illegal lexeme "'+n+'" for mode "'+(C.cN||"<unnamed>")+'"');return E+=n,n.length||1}var y=b(e);if(!y)throw new Error('Unknown language: "'+e+'"');c(y);var x,C=i||y,S={},k="";for(x=C;x!=y;x=x.parent)x.cN&&(k=h(x.cN,"",!0)+k);var E="",I=0;try{for(var T,M,j=0;C.t.lastIndex=j,T=C.t.exec(n),T;)M=v(n.substr(j,T.index-j),T[0]),j=T.index+M;for(v(n.substr(j)),x=C;x.parent;x=x.parent)x.cN&&(k+="</span>");return{r:I,value:k,language:e,top:C}}catch(F){if(-1!=F.message.indexOf("Illegal"))return{r:0,value:t(n)};throw F}}function d(e,n){n=n||w.languages||Object.keys(N);var r={r:0,value:t(e)},o=r;return n.forEach(function(t){if(b(t)){var n=u(t,e,!1);n.language=t,n.r>o.r&&(o=n),n.r>r.r&&(o=r,r=n)}}),o.language&&(r.second_best=o),r}function p(e){return w.tabReplace&&(e=e.replace(/^((<[^>]+>|\t)+)/gm,function(e,t){return t.replace(/\t/g,w.tabReplace)})),w.useBR&&(e=e.replace(/\n/g,"<br>")),e}function h(e,t,n){var r=t?x[t]:n,o=[e.trim()];return e.match(/\bhljs\b/)||o.push("hljs"),-1===e.indexOf(r)&&o.push(r),o.join(" ").trim()}function f(e){var t=i(e);if(!o(t)){var n;w.useBR?(n=document.createElementNS("http://www.w3.org/1999/xhtml","div"),n.innerHTML=e.innerHTML.replace(/\n/g,"").replace(/<br[ \/]*>/g,"\n")):n=e;var r=n.textContent,s=t?u(t,r,!0):d(r),c=a(n);if(c.length){var f=document.createElementNS("http://www.w3.org/1999/xhtml","div");f.innerHTML=s.value,s.value=l(c,a(f),r)}s.value=p(s.value),e.innerHTML=s.value,e.className=h(e.className,t,s.language),e.result={language:s.language,re:s.r},s.second_best&&(e.second_best={language:s.second_best.language,re:s.second_best.r})}}function m(e){w=s(w,e)}function g(){if(!g.called){g.called=!0;var e=document.querySelectorAll("pre code");Array.prototype.forEach.call(e,f)}}function _(){addEventListener("DOMContentLoaded",g,!1),addEventListener("load",g,!1)}function v(t,n){var r=N[t]=n(e);r.aliases&&r.aliases.forEach(function(e){x[e]=t})}function y(){return Object.keys(N)}function b(e){return e=(e||"").toLowerCase(),N[e]||N[x[e]]}var w={classPrefix:"hljs-",tabReplace:null,useBR:!1,languages:void 0},N={},x={};return e.highlight=u,e.highlightAuto=d,e.fixMarkup=p,e.highlightBlock=f,e.configure=m,e.initHighlighting=g,e.initHighlightingOnLoad=_,e.registerLanguage=v,e.listLanguages=y,e.getLanguage=b,e.inherit=s,e.IR="[a-zA-Z]\\w*",e.UIR="[a-zA-Z_]\\w*",e.NR="\\b\\d+(\\.\\d+)?",e.CNR="(\\b0[xX][a-fA-F0-9]+|(\\b\\d+(\\.\\d*)?|\\.\\d+)([eE][-+]?\\d+)?)",e.BNR="\\b(0b[01]+)",e.RSR="!|!=|!==|%|%=|&|&&|&=|\\*|\\*=|\\+|\\+=|,|-|-=|/=|/|:|;|<<|<<=|<=|<|===|==|=|>>>=|>>=|>=|>>>|>>|>|\\?|\\[|\\{|\\(|\\^|\\^=|\\||\\|=|\\|\\||~",e.BE={b:"\\\\[\\s\\S]",r:0},e.ASM={cN:"string",b:"'",e:"'",i:"\\n",c:[e.BE]},e.QSM={cN:"string",b:'"',e:'"',i:"\\n",c:[e.BE]},e.PWM={b:/\b(a|an|the|are|I|I'm|isn't|don't|doesn't|won't|but|just|should|pretty|simply|enough|gonna|going|wtf|so|such|will|you|your|like)\b/},e.C=function(t,n,r){var o=e.inherit({cN:"comment",b:t,e:n,c:[]},r||{});return o.c.push(e.PWM),o.c.push({cN:"doctag",b:"(?:TODO|FIXME|NOTE|BUG|XXX):",r:0}),o},e.CLCM=e.C("//","$"),e.CBCM=e.C("/\\*","\\*/"),e.HCM=e.C("#","$"),e.NM={cN:"number",b:e.NR,r:0},e.CNM={cN:"number",b:e.CNR,r:0},e.BNM={cN:"number",b:e.BNR,r:0},e.CSSNM={cN:"number",b:e.NR+"(%|em|ex|ch|rem|vw|vh|vmin|vmax|cm|mm|in|pt|pc|px|deg|grad|rad|turn|s|ms|Hz|kHz|dpi|dpcm|dppx)?",r:0},e.RM={cN:"regexp",b:/\//,e:/\/[gimuy]*/,i:/\n/,c:[e.BE,{b:/\[/,e:/\]/,r:0,c:[e.BE]}]},e.TM={cN:"title",b:e.IR,r:0},e.UTM={cN:"title",b:e.UIR,r:0},e}),hljs.registerLanguage("nginx",function(e){var t={cN:"variable",v:[{b:/\$\d+/},{b:/\$\{/,e:/}/},{b:"[\\$\\@]"+e.UIR}]},n={eW:!0,l:"[a-z/_]+",k:{built_in:"on off yes no true false none blocked debug info notice warn error crit select break last permanent redirect kqueue rtsig epoll poll /dev/poll"},r:0,i:"=>",c:[e.HCM,{cN:"string",c:[e.BE,t],v:[{b:/"/,e:/"/},{b:/'/,e:/'/}]},{cN:"url",b:"([a-z]+):/",e:"\\s",eW:!0,eE:!0,c:[t]},{cN:"regexp",c:[e.BE,t],v:[{b:"\\s\\^",e:"\\s|{|;",rE:!0},{b:"~\\*?\\s+",e:"\\s|{|;",rE:!0},{b:"\\*(\\.[a-z\\-]+)+"},{b:"([a-z\\-]+\\.)+\\*"}]},{cN:"number",b:"\\b\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}\\.\\d{1,3}(:\\d{1,5})?\\b"},{cN:"number",b:"\\b\\d+[kKmMgGdshdwy]*\\b",r:0},t]};return{aliases:["nginxconf"],c:[e.HCM,{b:e.UIR+"\\s",e:";|{",rB:!0,c:[{cN:"title",b:e.UIR,starts:n}],r:0}],i:"[^\\s\\}]"}}),hljs.registerLanguage("makefile",function(e){var t={cN:"variable",b:/\$\(/,e:/\)/,c:[e.BE]};return{aliases:["mk","mak"],c:[e.HCM,{b:/^\w+\s*\W*=/,rB:!0,r:0,starts:{cN:"constant",e:/\s*\W*=/,eE:!0,starts:{e:/$/,r:0,c:[t]}}},{cN:"title",b:/^[\w]+:\s*$/},{cN:"phony",b:/^\.PHONY:/,e:/$/,k:".PHONY",l:/[\.\w]+/},{b:/^\t+/,e:/$/,r:0,c:[e.QSM,t]}]}}),hljs.registerLanguage("http",function(e){return{aliases:["https"],i:"\\S",c:[{cN:"status",b:"^HTTP/[0-9\\.]+",e:"$",c:[{cN:"number",b:"\\b\\d{3}\\b"}]},{cN:"request",b:"^[A-Z]+ (.*?) HTTP/[0-9\\.]+$",rB:!0,e:"$",c:[{cN:"string",b:" ",e:" ",eB:!0,eE:!0}]},{cN:"attribute",b:"^\\w",e:": ",eE:!0,i:"\\n|\\s|=",starts:{cN:"string",e:"$"}},{b:"\\n\\n",starts:{sL:[],eW:!0}}]}}),hljs.registerLanguage("json",function(e){var t={literal:"true false null"},n=[e.QSM,e.CNM],r={cN:"value",e:",",eW:!0,eE:!0,c:n,k:t},o={b:"{",e:"}",c:[{cN:"attribute",b:'\\s*"',e:'"\\s*:\\s*',eB:!0,eE:!0,c:[e.BE],i:"\\n",starts:r}],i:"\\S"},i={b:"\\[",e:"\\]",c:[e.inherit(r,{cN:null})],i:"\\S"};return n.splice(n.length,0,o,i),{c:n,k:t,i:"\\S"}}),hljs.registerLanguage("css",function(e){var t="[a-zA-Z-][a-zA-Z0-9_-]*",n={cN:"function",b:t+"\\(",rB:!0,eE:!0,e:"\\("},r={cN:"rule",b:/[A-Z\_\.\-]+\s*:/,rB:!0,e:";",eW:!0,c:[{cN:"attribute",b:/\S/,e:":",eE:!0,starts:{cN:"value",eW:!0,eE:!0,c:[n,e.CSSNM,e.QSM,e.ASM,e.CBCM,{cN:"hexcolor",b:"#[0-9A-Fa-f]+"},{cN:"important",b:"!important"}]}}]};return{cI:!0,i:/[=\/|'\$]/,c:[e.CBCM,{cN:"id",b:/\#[A-Za-z0-9_-]+/},{cN:"class",b:/\.[A-Za-z0-9_-]+/},{cN:"attr_selector",b:/\[/,e:/\]/,i:"$"},{cN:"pseudo",b:/:(:)?[a-zA-Z0-9\_\-\+\(\)"']+/},{cN:"at_rule",b:"@(font-face|page)",l:"[a-z-]+",k:"font-face page"},{cN:"at_rule",b:"@",e:"[{;]",c:[{cN:"keyword",b:/\S+/},{b:/\s/,eW:!0,eE:!0,r:0,c:[n,e.ASM,e.QSM,e.CSSNM]}]},{cN:"tag",b:t,r:0},{cN:"rules",b:"{",e:"}",i:/\S/,c:[e.CBCM,r]}]}}),hljs.registerLanguage("bash",function(e){var t={cN:"variable",v:[{b:/\$[\w\d#@][\w\d_]*/},{b:/\$\{(.*?)}/}]},n={cN:"string",b:/"/,e:/"/,c:[e.BE,t,{cN:"variable",b:/\$\(/,e:/\)/,c:[e.BE]}]},r={cN:"string",b:/'/,e:/'/};return{aliases:["sh","zsh"],l:/-?[a-z\.]+/,k:{keyword:"if then else elif fi for while in do done case esac function",literal:"true false",built_in:"break cd continue eval exec exit export getopts hash pwd readonly return shift test times trap umask unset alias bind builtin caller command declare echo enable help let local logout mapfile printf read readarray source type typeset ulimit unalias set shopt autoload bg bindkey bye cap chdir clone comparguments compcall compctl compdescribe compfiles compgroups compquote comptags comptry compvalues dirs disable disown echotc echoti emulate fc fg float functions getcap getln history integer jobs kill limit log noglob popd print pushd pushln rehash sched setcap setopt stat suspend ttyctl unfunction unhash unlimit unsetopt vared wait whence where which zcompile zformat zftp zle zmodload zparseopts zprof zpty zregexparse zsocket zstyle ztcp",operator:"-ne -eq -lt -gt -f -d -e -s -l -a"},c:[{cN:"shebang",b:/^#![^\n]+sh\s*$/,r:10},{cN:"function",b:/\w[\w\d_]*\s*\(\s*\)\s*\{/,rB:!0,c:[e.inherit(e.TM,{b:/\w[\w\d_]*/})],r:0},e.HCM,e.NM,n,r,t]}}),hljs.registerLanguage("xml",function(e){var t="[A-Za-z0-9\\._:-]+",n={b:/<\?(php)?(?!\w)/,e:/\?>/,sL:"php"},r={eW:!0,i:/</,r:0,c:[n,{cN:"attribute",b:t,r:0},{b:"=",r:0,c:[{cN:"value",c:[n],v:[{b:/"/,e:/"/},{b:/'/,e:/'/},{b:/[^\s\/>]+/}]}]}]};return{aliases:["html","xhtml","rss","atom","xsl","plist"],cI:!0,c:[{cN:"doctype",b:"<!DOCTYPE",e:">",r:10,c:[{b:"\\[",e:"\\]"}]},e.C("<!--","-->",{r:10}),{cN:"cdata",b:"<\\!\\[CDATA\\[",e:"\\]\\]>",r:10},{cN:"tag",b:"<style(?=\\s|>|$)",e:">",k:{title:"style"},c:[r],starts:{e:"</style>",rE:!0,sL:"css"}},{cN:"tag",b:"<script(?=\\s|>|$)",e:">",k:{title:"script"},c:[r],starts:{e:"</script>",rE:!0,sL:["actionscript","javascript","handlebars"]}},n,{cN:"pi",b:/<\?\w+/,e:/\?>/,r:10},{cN:"tag",b:"</?",e:"/?>",c:[{cN:"title",b:/[^ \/><\n\t]+/,r:0},r]}]}}),hljs.registerLanguage("markdown",function(e){return{aliases:["md","mkdown","mkd"],c:[{cN:"header",v:[{b:"^#{1,6}",e:"$"},{b:"^.+?\\n[=-]{2,}$"}]},{b:"<",e:">",sL:"xml",r:0},{cN:"bullet",b:"^([*+-]|(\\d+\\.))\\s+"},{cN:"strong",b:"[*_]{2}.+?[*_]{2}"},{cN:"emphasis",v:[{b:"\\*.+?\\*"},{b:"_.+?_",r:0}]},{cN:"blockquote",b:"^>\\s+",e:"$"},{cN:"code",v:[{b:"`.+?`"},{b:"^( {4}|   )",e:"$",r:0}]},{cN:"horizontal_rule",b:"^[-\\*]{3,}",e:"$"},{b:"\\[.+?\\][\\(\\[].*?[\\)\\]]",rB:!0,c:[{cN:"link_label",b:"\\[",e:"\\]",eB:!0,rE:!0,r:0},{cN:"link_url",b:"\\]\\(",e:"\\)",eB:!0,eE:!0},{cN:"link_reference",b:"\\]\\[",e:"\\]",eB:!0,eE:!0}],r:10},{b:"^\\[.+\\]:",rB:!0,c:[{cN:"link_reference",b:"\\[",e:"\\]:",eB:!0,eE:!0,starts:{cN:"link_url",e:"$"}}]}]}}),hljs.registerLanguage("php",function(e){var t={cN:"variable",b:"\\$+[a-zA-Z_-ÿ][a-zA-Z0-9_-ÿ]*"},n={cN:"preprocessor",b:/<\?(php)?|\?>/},r={cN:"string",c:[e.BE,n],v:[{b:'b"',e:'"'},{b:"b'",e:"'"},e.inherit(e.ASM,{i:null}),e.inherit(e.QSM,{i:null})]},o={v:[e.BNM,e.CNM]};return{aliases:["php3","php4","php5","php6"],cI:!0,k:"and include_once list abstract global private echo interface as static endswitch array null if endwhile or const for endforeach self var while isset public protected exit foreach throw elseif include __FILE__ empty require_once do xor return parent clone use __CLASS__ __LINE__ else break print eval new catch __METHOD__ case exception default die require __FUNCTION__ enddeclare final try switch continue endfor endif declare unset true false trait goto instanceof insteadof __DIR__ __NAMESPACE__ yield finally",c:[e.CLCM,e.HCM,e.C("/\\*","\\*/",{c:[{cN:"doctag",b:"@[A-Za-z]+"},n]}),e.C("__halt_compiler.+?;",!1,{eW:!0,k:"__halt_compiler",l:e.UIR}),{cN:"string",b:/<<<['"]?\w+['"]?$/,e:/^\w+;?$/,c:[e.BE,{cN:"subst",v:[{b:/\$\w+/},{b:/\{\$/,e:/\}/}]}]},n,t,{b:/(::|->)+[a-zA-Z_\x7f-\xff][a-zA-Z0-9_\x7f-\xff]*/},{cN:"function",bK:"function",e:/[;{]/,eE:!0,i:"\\$|\\[|%",c:[e.UTM,{cN:"params",b:"\\(",e:"\\)",c:["self",t,e.CBCM,r,o]}]},{cN:"class",bK:"class interface",e:"{",eE:!0,i:/[:\(\$"]/,c:[{bK:"extends implements"},e.UTM]},{bK:"namespace",e:";",i:/[\.']/,c:[e.UTM]},{bK:"use",e:";",c:[e.UTM]},{b:"=>"},r,o]}}),hljs.registerLanguage("javascript",function(e){return{aliases:["js"],k:{keyword:"in of if for while finally var new function do return void else break catch instanceof with throw case default try this switch continue typeof delete let yield const export super debugger as async await",literal:"true false null undefined NaN Infinity",built_in:"eval isFinite isNaN parseFloat parseInt decodeURI decodeURIComponent encodeURI encodeURIComponent escape unescape Object Function Boolean Error EvalError InternalError RangeError ReferenceError StopIteration SyntaxError TypeError URIError Number Math Date String RegExp Array Float32Array Float64Array Int16Array Int32Array Int8Array Uint16Array Uint32Array Uint8Array Uint8ClampedArray ArrayBuffer DataView JSON Intl arguments require module console window document Symbol Set Map WeakSet WeakMap Proxy Reflect Promise"},c:[{cN:"pi",r:10,b:/^\s*['"]use (strict|asm)['"]/},e.ASM,e.QSM,{cN:"string",b:"`",e:"`",c:[e.BE,{cN:"subst",b:"\\$\\{",e:"\\}"}]},e.CLCM,e.CBCM,{cN:"number",v:[{b:"\\b(0[bB][01]+)"},{b:"\\b(0[oO][0-7]+)"},{b:e.CNR}],r:0},{b:"("+e.RSR+"|\\b(case|return|throw)\\b)\\s*",k:"return throw case",c:[e.CLCM,e.CBCM,e.RM,{b:/</,e:/>\s*[);\]]/,r:0,sL:"xml"}],r:0},{cN:"function",bK:"function",e:/\{/,eE:!0,c:[e.inherit(e.TM,{b:/[A-Za-z$_][0-9A-Za-z$_]*/}),{cN:"params",b:/\(/,e:/\)/,eB:!0,eE:!0,c:[e.CLCM,e.CBCM]}],i:/\[|%/},{b:/\$[(.]/},{b:"\\."+e.IR,r:0},{bK:"import",e:"[;$]",k:"import from as",c:[e.ASM,e.QSM]},{cN:"class",bK:"class",e:/[{;=]/,eE:!0,i:/[:"\[\]]/,c:[{bK:"extends"},e.UTM]}],i:/#/}}),hljs.registerLanguage("ini",function(e){var t={cN:"string",c:[e.BE],v:[{b:"'''",e:"'''",r:10},{b:'"""',e:'"""',r:10},{b:'"',e:'"'},{b:"'",e:"'"}]};return{aliases:["toml"],cI:!0,i:/\S/,c:[e.C(";","$"),e.HCM,{cN:"title",b:/^\s*\[+/,e:/\]+/},{cN:"setting",b:/^[a-z0-9\[\]_-]+\s*=\s*/,e:"$",c:[{cN:"value",eW:!0,k:"on off true false yes no",c:[{cN:"variable",v:[{b:/\$[\w\d"][\w\d_]*/},{b:/\$\{(.*?)}/}]},t,{cN:"number",b:/([\+\-]+)?[\d]+_[\d_]+/},e.NM],r:0}]}]}}),hljs.registerLanguage("apache",function(e){var t={cN:"number",b:"[\\$%]\\d+"};return{aliases:["apacheconf"],cI:!0,c:[e.HCM,{cN:"tag",b:"</?",e:">"},{cN:"keyword",b:/\w+/,r:0,k:{common:"order deny allow setenv rewriterule rewriteengine rewritecond documentroot sethandler errordocument loadmodule options header listen serverroot servername"},starts:{e:/$/,r:0,k:{literal:"on off all"},c:[{cN:"sqbracket",b:"\\s\\[",e:"\\]$"},{cN:"cbracket",b:"[\\$%]\\{",e:"\\}",c:["self",t]},t,e.QSM]}}],i:/\S/}}),hljs.registerLanguage("sql",function(e){var t=e.C("--","$");return{cI:!0,i:/[<>{}*]/,c:[{cN:"operator",bK:"begin end start commit rollback savepoint lock alter create drop rename call delete do handler insert load replace select truncate update set show pragma grant merge describe use explain help declare prepare execute deallocate release unlock purge reset change stop analyze cache flush optimize repair kill install uninstall checksum restore check backup revoke",e:/;/,eW:!0,k:{keyword:"abort abs absolute acc acce accep accept access accessed accessible account acos action activate add addtime admin administer advanced advise aes_decrypt aes_encrypt after agent aggregate ali alia alias allocate allow alter always analyze ancillary and any anydata anydataset anyschema anytype apply archive archived archivelog are as asc ascii asin assembly assertion associate asynchronous at atan atn2 attr attri attrib attribu attribut attribute attributes audit authenticated authentication authid authors auto autoallocate autodblink autoextend automatic availability avg backup badfile basicfile before begin beginning benchmark between bfile bfile_base big bigfile bin binary_double binary_float binlog bit_and bit_count bit_length bit_or bit_xor bitmap blob_base block blocksize body both bound buffer_cache buffer_pool build bulk by byte byteordermark bytes c cache caching call calling cancel capacity cascade cascaded case cast catalog category ceil ceiling chain change changed char_base char_length character_length characters characterset charindex charset charsetform charsetid check checksum checksum_agg child choose chr chunk class cleanup clear client clob clob_base clone close cluster_id cluster_probability cluster_set clustering coalesce coercibility col collate collation collect colu colum column column_value columns columns_updated comment commit compact compatibility compiled complete composite_limit compound compress compute concat concat_ws concurrent confirm conn connec connect connect_by_iscycle connect_by_isleaf connect_by_root connect_time connection consider consistent constant constraint constraints constructor container content contents context contributors controlfile conv convert convert_tz corr corr_k corr_s corresponding corruption cos cost count count_big counted covar_pop covar_samp cpu_per_call cpu_per_session crc32 create creation critical cross cube cume_dist curdate current current_date current_time current_timestamp current_user cursor curtime customdatum cycle d data database databases datafile datafiles datalength date_add date_cache date_format date_sub dateadd datediff datefromparts datename datepart datetime2fromparts day day_to_second dayname dayofmonth dayofweek dayofyear days db_role_change dbtimezone ddl deallocate declare decode decompose decrement decrypt deduplicate def defa defau defaul default defaults deferred defi defin define degrees delayed delegate delete delete_all delimited demand dense_rank depth dequeue des_decrypt des_encrypt des_key_file desc descr descri describ describe descriptor deterministic diagnostics difference dimension direct_load directory disable disable_all disallow disassociate discardfile disconnect diskgroup distinct distinctrow distribute distributed div do document domain dotnet double downgrade drop dumpfile duplicate duration e each edition editionable editions element ellipsis else elsif elt empty enable enable_all enclosed encode encoding encrypt end end-exec endian enforced engine engines enqueue enterprise entityescaping eomonth error errors escaped evalname evaluate event eventdata events except exception exceptions exchange exclude excluding execu execut execute exempt exists exit exp expire explain export export_set extended extent external external_1 external_2 externally extract f failed failed_login_attempts failover failure far fast feature_set feature_value fetch field fields file file_name_convert filesystem_like_logging final finish first first_value fixed flash_cache flashback floor flush following follows for forall force form forma format found found_rows freelist freelists freepools fresh from from_base64 from_days ftp full function g general generated get get_format get_lock getdate getutcdate global global_name globally go goto grant grants greatest group group_concat group_id grouping grouping_id groups gtid_subtract guarantee guard handler hash hashkeys having hea head headi headin heading heap help hex hierarchy high high_priority hosts hour http i id ident_current ident_incr ident_seed identified identity idle_time if ifnull ignore iif ilike ilm immediate import in include including increment index indexes indexing indextype indicator indices inet6_aton inet6_ntoa inet_aton inet_ntoa infile initial initialized initially initrans inmemory inner innodb input insert install instance instantiable instr interface interleaved intersect into invalidate invisible is is_free_lock is_ipv4 is_ipv4_compat is_not is_not_null is_used_lock isdate isnull isolation iterate java join json json_exists k keep keep_duplicates key keys kill l language large last last_day last_insert_id last_value lax lcase lead leading least leaves left len lenght length less level levels library like like2 like4 likec limit lines link list listagg little ln load load_file lob lobs local localtime localtimestamp locate locator lock locked log log10 log2 logfile logfiles logging logical logical_reads_per_call logoff logon logs long loop low low_priority lower lpad lrtrim ltrim m main make_set makedate maketime managed management manual map mapping mask master master_pos_wait match matched materialized max maxextents maximize maxinstances maxlen maxlogfiles maxloghistory maxlogmembers maxsize maxtrans md5 measures median medium member memcompress memory merge microsecond mid migration min minextents minimum mining minus minute minvalue missing mod mode model modification modify module monitoring month months mount move movement multiset mutex n name name_const names nan national native natural nav nchar nclob nested never new newline next nextval no no_write_to_binlog noarchivelog noaudit nobadfile nocheck nocompress nocopy nocycle nodelay nodiscardfile noentityescaping noguarantee nokeep nologfile nomapping nomaxvalue nominimize nominvalue nomonitoring none noneditionable nonschema noorder nopr nopro noprom nopromp noprompt norely noresetlogs noreverse normal norowdependencies noschemacheck noswitch not nothing notice notrim novalidate now nowait nth_value nullif nulls num numb numbe nvarchar nvarchar2 object ocicoll ocidate ocidatetime ociduration ociinterval ociloblocator ocinumber ociref ocirefcursor ocirowid ocistring ocitype oct octet_length of off offline offset oid oidindex old on online only opaque open operations operator optimal optimize option optionally or oracle oracle_date oradata ord ordaudio orddicom orddoc order ordimage ordinality ordvideo organization orlany orlvary out outer outfile outline output over overflow overriding p package pad parallel parallel_enable parameters parent parse partial partition partitions pascal passing password password_grace_time password_lock_time password_reuse_max password_reuse_time password_verify_function patch path patindex pctincrease pctthreshold pctused pctversion percent percent_rank percentile_cont percentile_disc performance period period_add period_diff permanent physical pi pipe pipelined pivot pluggable plugin policy position post_transaction pow power pragma prebuilt precedes preceding precision prediction prediction_cost prediction_details prediction_probability prediction_set prepare present preserve prior priority private private_sga privileges procedural procedure procedure_analyze processlist profiles project prompt protection public publishingservername purge quarter query quick quiesce quota quotename radians raise rand range rank raw read reads readsize rebuild record records recover recovery recursive recycle redo reduced ref reference referenced references referencing refresh regexp_like register regr_avgx regr_avgy regr_count regr_intercept regr_r2 regr_slope regr_sxx regr_sxy reject rekey relational relative relaylog release release_lock relies_on relocate rely rem remainder rename repair repeat replace replicate replication required reset resetlogs resize resource respect restore restricted result result_cache resumable resume retention return returning returns reuse reverse revoke right rlike role roles rollback rolling rollup round row row_count rowdependencies rowid rownum rows rtrim rules safe salt sample save savepoint sb1 sb2 sb4 scan schema schemacheck scn scope scroll sdo_georaster sdo_topo_geometry search sec_to_time second section securefile security seed segment select self sequence sequential serializable server servererror session session_user sessions_per_user set sets settings sha sha1 sha2 share shared shared_pool short show shrink shutdown si_averagecolor si_colorhistogram si_featurelist si_positionalcolor si_stillimage si_texture siblings sid sign sin size size_t sizes skip slave sleep smalldatetimefromparts smallfile snapshot some soname sort soundex source space sparse spfile split sql sql_big_result sql_buffer_result sql_cache sql_calc_found_rows sql_small_result sql_variant_property sqlcode sqldata sqlerror sqlname sqlstate sqrt square standalone standby start starting startup statement static statistics stats_binomial_test stats_crosstab stats_ks_test stats_mode stats_mw_test stats_one_way_anova stats_t_test_ stats_t_test_indep stats_t_test_one stats_t_test_paired stats_wsr_test status std stddev stddev_pop stddev_samp stdev stop storage store stored str str_to_date straight_join strcmp strict string struct stuff style subdate subpartition subpartitions substitutable substr substring subtime subtring_index subtype success sum suspend switch switchoffset switchover sync synchronous synonym sys sys_xmlagg sysasm sysaux sysdate sysdatetimeoffset sysdba sysoper system system_user sysutcdatetime t table tables tablespace tan tdo template temporary terminated tertiary_weights test than then thread through tier ties time time_format time_zone timediff timefromparts timeout timestamp timestampadd timestampdiff timezone_abbr timezone_minute timezone_region to to_base64 to_date to_days to_seconds todatetimeoffset trace tracking transaction transactional translate translation treat trigger trigger_nestlevel triggers trim truncate try_cast try_convert try_parse type ub1 ub2 ub4 ucase unarchived unbounded uncompress under undo unhex unicode uniform uninstall union unique unix_timestamp unknown unlimited unlock unpivot unrecoverable unsafe unsigned until untrusted unusable unused update updated upgrade upped upper upsert url urowid usable usage use use_stored_outlines user user_data user_resources users using utc_date utc_timestamp uuid uuid_short validate validate_password_strength validation valist value values var var_samp varcharc vari varia variab variabl variable variables variance varp varraw varrawc varray verify version versions view virtual visible void wait wallet warning warnings week weekday weekofyear wellformed when whene whenev wheneve whenever where while whitespace with within without work wrapped xdb xml xmlagg xmlattributes xmlcast xmlcolattval xmlelement xmlexists xmlforest xmlindex xmlnamespaces xmlpi xmlquery xmlroot xmlschema xmlserialize xmltable xmltype xor year year_to_month years yearweek",
+literal:"true false null",built_in:"array bigint binary bit blob boolean char character date dec decimal float int int8 integer interval number numeric real record serial serial8 smallint text varchar varying void"},c:[{cN:"string",b:"'",e:"'",c:[e.BE,{b:"''"}]},{cN:"string",b:'"',e:'"',c:[e.BE,{b:'""'}]},{cN:"string",b:"`",e:"`",c:[e.BE]},e.CNM,e.CBCM,t]},e.CBCM,t]}}),hljs.registerLanguage("twig",function(e){var t={cN:"params",b:"\\(",e:"\\)"},n="attribute block constant cycle date dump include max min parent random range source template_from_string",r={cN:"function",bK:n,r:0,c:[t]},o={cN:"filter",b:/\|[A-Za-z_]+:?/,k:"abs batch capitalize convert_encoding date date_modify default escape first format join json_encode keys last length lower merge nl2br number_format raw replace reverse round slice sort split striptags title trim upper url_encode",c:[r]},i="autoescape block do embed extends filter flush for if import include macro sandbox set spaceless use verbatim";return i=i+" "+i.split(" ").map(function(e){return"end"+e}).join(" "),{aliases:["craftcms"],cI:!0,sL:"xml",c:[e.C(/\{#/,/#}/),{cN:"template_tag",b:/\{%/,e:/%}/,k:i,c:[o,r]},{cN:"variable",b:/\{\{/,e:/}}/,c:[o,r]}]}}),!function(e){"function"==typeof define&&define.amd?define(["jquery"],e):e("object"==typeof exports?require("jquery"):window.jQuery||window.Zepto)}(function(e){var t,n,r,o,i,s,a="Close",l="BeforeClose",c="AfterClose",u="BeforeAppend",d="MarkupParse",p="Open",h="Change",f="mfp",m="."+f,g="mfp-ready",_="mfp-removing",v="mfp-prevent-close",y=function(){},b=!!window.jQuery,w=e(window),N=function(e,n){t.ev.on(f+e+m,n)},x=function(t,n,r,o){var i=document.createElement("div");return i.className="mfp-"+t,r&&(i.innerHTML=r),o?n&&n.appendChild(i):(i=e(i),n&&i.appendTo(n)),i},C=function(n,r){t.ev.triggerHandler(f+n,r),t.st.callbacks&&(n=n.charAt(0).toLowerCase()+n.slice(1),t.st.callbacks[n]&&t.st.callbacks[n].apply(t,e.isArray(r)?r:[r]))},S=function(n){return n===s&&t.currTemplate.closeBtn||(t.currTemplate.closeBtn=e(t.st.closeMarkup.replace("%title%",t.st.tClose)),s=n),t.currTemplate.closeBtn},k=function(){e.magnificPopup.instance||(t=new y,t.init(),e.magnificPopup.instance=t)},E=function(){var e=document.createElement("p").style,t=["ms","O","Moz","Webkit"];if(void 0!==e.transition)return!0;for(;t.length;)if(t.pop()+"Transition"in e)return!0;return!1};y.prototype={constructor:y,init:function(){var n=navigator.appVersion;t.isIE7=-1!==n.indexOf("MSIE 7."),t.isIE8=-1!==n.indexOf("MSIE 8."),t.isLowIE=t.isIE7||t.isIE8,t.isAndroid=/android/gi.test(n),t.isIOS=/iphone|ipad|ipod/gi.test(n),t.supportsTransition=E(),t.probablyMobile=t.isAndroid||t.isIOS||/(Opera Mini)|Kindle|webOS|BlackBerry|(Opera Mobi)|(Windows Phone)|IEMobile/i.test(navigator.userAgent),r=e(document),t.popupsCache={}},open:function(n){var o;if(n.isObj===!1){t.items=n.items.toArray(),t.index=0;var s,a=n.items;for(o=0;o<a.length;o++)if(s=a[o],s.parsed&&(s=s.el[0]),s===n.el[0]){t.index=o;break}}else t.items=e.isArray(n.items)?n.items:[n.items],t.index=n.index||0;if(t.isOpen)return void t.updateItemHTML();t.types=[],i="",n.mainEl&&n.mainEl.length?t.ev=n.mainEl.eq(0):t.ev=r,n.key?(t.popupsCache[n.key]||(t.popupsCache[n.key]={}),t.currTemplate=t.popupsCache[n.key]):t.currTemplate={},t.st=e.extend(!0,{},e.magnificPopup.defaults,n),t.fixedContentPos="auto"===t.st.fixedContentPos?!t.probablyMobile:t.st.fixedContentPos,t.st.modal&&(t.st.closeOnContentClick=!1,t.st.closeOnBgClick=!1,t.st.showCloseBtn=!1,t.st.enableEscapeKey=!1),t.bgOverlay||(t.bgOverlay=x("bg").on("click"+m,function(){t.close()}),t.wrap=x("wrap").attr("tabindex",-1).on("click"+m,function(e){t._checkIfClose(e.target)&&t.close()}),t.container=x("container",t.wrap)),t.contentContainer=x("content"),t.st.preloader&&(t.preloader=x("preloader",t.container,t.st.tLoading));var l=e.magnificPopup.modules;for(o=0;o<l.length;o++){var c=l[o];c=c.charAt(0).toUpperCase()+c.slice(1),t["init"+c].call(t)}C("BeforeOpen"),t.st.showCloseBtn&&(t.st.closeBtnInside?(N(d,function(e,t,n,r){n.close_replaceWith=S(r.type)}),i+=" mfp-close-btn-in"):t.wrap.append(S())),t.st.alignTop&&(i+=" mfp-align-top"),t.fixedContentPos?t.wrap.css({overflow:t.st.overflowY,overflowX:"hidden",overflowY:t.st.overflowY}):t.wrap.css({top:w.scrollTop(),position:"absolute"}),(t.st.fixedBgPos===!1||"auto"===t.st.fixedBgPos&&!t.fixedContentPos)&&t.bgOverlay.css({height:r.height(),position:"absolute"}),t.st.enableEscapeKey&&r.on("keyup"+m,function(e){27===e.keyCode&&t.close()}),w.on("resize"+m,function(){t.updateSize()}),t.st.closeOnContentClick||(i+=" mfp-auto-cursor"),i&&t.wrap.addClass(i);var u=t.wH=w.height(),h={};if(t.fixedContentPos&&t._hasScrollBar(u)){var f=t._getScrollbarSize();f&&(h.marginRight=f)}t.fixedContentPos&&(t.isIE7?e("body, html").css("overflow","hidden"):h.overflow="hidden");var _=t.st.mainClass;return t.isIE7&&(_+=" mfp-ie7"),_&&t._addClassToMFP(_),t.updateItemHTML(),C("BuildControls"),e("html").css(h),t.bgOverlay.add(t.wrap).prependTo(t.st.prependTo||e(document.body)),t._lastFocusedEl=document.activeElement,setTimeout(function(){t.content?(t._addClassToMFP(g),t._setFocus()):t.bgOverlay.addClass(g),r.on("focusin"+m,t._onFocusIn)},16),t.isOpen=!0,t.updateSize(u),C(p),n},close:function(){t.isOpen&&(C(l),t.isOpen=!1,t.st.removalDelay&&!t.isLowIE&&t.supportsTransition?(t._addClassToMFP(_),setTimeout(function(){t._close()},t.st.removalDelay)):t._close())},_close:function(){C(a);var n=_+" "+g+" ";if(t.bgOverlay.detach(),t.wrap.detach(),t.container.empty(),t.st.mainClass&&(n+=t.st.mainClass+" "),t._removeClassFromMFP(n),t.fixedContentPos){var o={marginRight:""};t.isIE7?e("body, html").css("overflow",""):o.overflow="",e("html").css(o)}r.off("keyup"+m+" focusin"+m),t.ev.off(m),t.wrap.attr("class","mfp-wrap").removeAttr("style"),t.bgOverlay.attr("class","mfp-bg"),t.container.attr("class","mfp-container"),!t.st.showCloseBtn||t.st.closeBtnInside&&t.currTemplate[t.currItem.type]!==!0||t.currTemplate.closeBtn&&t.currTemplate.closeBtn.detach(),t._lastFocusedEl&&e(t._lastFocusedEl).focus(),t.currItem=null,t.content=null,t.currTemplate=null,t.prevHeight=0,C(c)},updateSize:function(e){if(t.isIOS){var n=document.documentElement.clientWidth/window.innerWidth,r=window.innerHeight*n;t.wrap.css("height",r),t.wH=r}else t.wH=e||w.height();t.fixedContentPos||t.wrap.css("height",t.wH),C("Resize")},updateItemHTML:function(){var n=t.items[t.index];t.contentContainer.detach(),t.content&&t.content.detach(),n.parsed||(n=t.parseEl(t.index));var r=n.type;if(C("BeforeChange",[t.currItem?t.currItem.type:"",r]),t.currItem=n,!t.currTemplate[r]){var i=t.st[r]?t.st[r].markup:!1;C("FirstMarkupParse",i),i?t.currTemplate[r]=e(i):t.currTemplate[r]=!0}o&&o!==n.type&&t.container.removeClass("mfp-"+o+"-holder");var s=t["get"+r.charAt(0).toUpperCase()+r.slice(1)](n,t.currTemplate[r]);t.appendContent(s,r),n.preloaded=!0,C(h,n),o=n.type,t.container.prepend(t.contentContainer),C("AfterChange")},appendContent:function(e,n){t.content=e,e?t.st.showCloseBtn&&t.st.closeBtnInside&&t.currTemplate[n]===!0?t.content.find(".mfp-close").length||t.content.append(S()):t.content=e:t.content="",C(u),t.container.addClass("mfp-"+n+"-holder"),t.contentContainer.append(t.content)},parseEl:function(n){var r,o=t.items[n];if(o.tagName?o={el:e(o)}:(r=o.type,o={data:o,src:o.src}),o.el){for(var i=t.types,s=0;s<i.length;s++)if(o.el.hasClass("mfp-"+i[s])){r=i[s];break}o.src=o.el.attr("data-mfp-src"),o.src||(o.src=o.el.attr("href"))}return o.type=r||t.st.type||"inline",o.index=n,o.parsed=!0,t.items[n]=o,C("ElementParse",o),t.items[n]},addGroup:function(e,n){var r=function(r){r.mfpEl=this,t._openClick(r,e,n)};n||(n={});var o="click.magnificPopup";n.mainEl=e,n.items?(n.isObj=!0,e.off(o).on(o,r)):(n.isObj=!1,n.delegate?e.off(o).on(o,n.delegate,r):(n.items=e,e.off(o).on(o,r)))},_openClick:function(n,r,o){var i=void 0!==o.midClick?o.midClick:e.magnificPopup.defaults.midClick;if(i||!(2===n.which||n.ctrlKey||n.metaKey||n.altKey||n.shiftKey)){var s=void 0!==o.disableOn?o.disableOn:e.magnificPopup.defaults.disableOn;if(s)if(e.isFunction(s)){if(!s.call(t))return!0}else if(w.width()<s)return!0;n.type&&(n.preventDefault(),t.isOpen&&n.stopPropagation()),o.el=e(n.mfpEl),o.delegate&&(o.items=r.find(o.delegate)),t.open(o)}},updateStatus:function(e,r){if(t.preloader){n!==e&&t.container.removeClass("mfp-s-"+n),r||"loading"!==e||(r=t.st.tLoading);var o={status:e,text:r};C("UpdateStatus",o),e=o.status,r=o.text,t.preloader.html(r),t.preloader.find("a").on("click",function(e){e.stopImmediatePropagation()}),t.container.addClass("mfp-s-"+e),n=e}},_checkIfClose:function(n){if(!e(n).hasClass(v)){var r=t.st.closeOnContentClick,o=t.st.closeOnBgClick;if(r&&o)return!0;if(!t.content||e(n).hasClass("mfp-close")||t.preloader&&n===t.preloader[0])return!0;if(n===t.content[0]||e.contains(t.content[0],n)){if(r)return!0}else if(o&&e.contains(document,n))return!0;return!1}},_addClassToMFP:function(e){t.bgOverlay.addClass(e),t.wrap.addClass(e)},_removeClassFromMFP:function(e){this.bgOverlay.removeClass(e),t.wrap.removeClass(e)},_hasScrollBar:function(e){return(t.isIE7?r.height():document.body.scrollHeight)>(e||w.height())},_setFocus:function(){(t.st.focus?t.content.find(t.st.focus).eq(0):t.wrap).focus()},_onFocusIn:function(n){return n.target===t.wrap[0]||e.contains(t.wrap[0],n.target)?void 0:(t._setFocus(),!1)},_parseMarkup:function(t,n,r){var o;r.data&&(n=e.extend(r.data,n)),C(d,[t,n,r]),e.each(n,function(e,n){if(void 0===n||n===!1)return!0;if(o=e.split("_"),o.length>1){var r=t.find(m+"-"+o[0]);if(r.length>0){var i=o[1];"replaceWith"===i?r[0]!==n[0]&&r.replaceWith(n):"img"===i?r.is("img")?r.attr("src",n):r.replaceWith('<img src="'+n+'" class="'+r.attr("class")+'" />'):r.attr(o[1],n)}}else t.find(m+"-"+e).html(n)})},_getScrollbarSize:function(){if(void 0===t.scrollbarSize){var e=document.createElement("div");e.style.cssText="width: 99px; height: 99px; overflow: scroll; position: absolute; top: -9999px;",document.body.appendChild(e),t.scrollbarSize=e.offsetWidth-e.clientWidth,document.body.removeChild(e)}return t.scrollbarSize}},e.magnificPopup={instance:null,proto:y.prototype,modules:[],open:function(t,n){return k(),t=t?e.extend(!0,{},t):{},t.isObj=!0,t.index=n||0,this.instance.open(t)},close:function(){return e.magnificPopup.instance&&e.magnificPopup.instance.close()},registerModule:function(t,n){n.options&&(e.magnificPopup.defaults[t]=n.options),e.extend(this.proto,n.proto),this.modules.push(t)},defaults:{disableOn:0,key:null,midClick:!1,mainClass:"",preloader:!0,focus:"",closeOnContentClick:!1,closeOnBgClick:!0,closeBtnInside:!0,showCloseBtn:!0,enableEscapeKey:!0,modal:!1,alignTop:!1,removalDelay:0,prependTo:null,fixedContentPos:"auto",fixedBgPos:"auto",overflowY:"auto",closeMarkup:'<button title="%title%" type="button" class="mfp-close">&#215;</button>',tClose:"Close (Esc)",tLoading:"Loading..."}},e.fn.magnificPopup=function(n){k();var r=e(this);if("string"==typeof n)if("open"===n){var o,i=b?r.data("magnificPopup"):r[0].magnificPopup,s=parseInt(arguments[1],10)||0;i.items?o=i.items[s]:(o=r,i.delegate&&(o=o.find(i.delegate)),o=o.eq(s)),t._openClick({mfpEl:o},r,i)}else t.isOpen&&t[n].apply(t,Array.prototype.slice.call(arguments,1));else n=e.extend(!0,{},n),b?r.data("magnificPopup",n):r[0].magnificPopup=n,t.addGroup(r,n);return r};var I,T,M,j="inline",F=function(){M&&(T.after(M.addClass(I)).detach(),M=null)};e.magnificPopup.registerModule(j,{options:{hiddenClass:"hide",markup:"",tNotFound:"Content not found"},proto:{initInline:function(){t.types.push(j),N(a+"."+j,function(){F()})},getInline:function(n,r){if(F(),n.src){var o=t.st.inline,i=e(n.src);if(i.length){var s=i[0].parentNode;s&&s.tagName&&(T||(I=o.hiddenClass,T=x(I),I="mfp-"+I),M=i.after(T).detach().removeClass(I)),t.updateStatus("ready")}else t.updateStatus("error",o.tNotFound),i=e("<div>");return n.inlineElement=i,i}return t.updateStatus("ready"),t._parseMarkup(r,{},n),r}}});var A,O="ajax",D=function(){A&&e(document.body).removeClass(A)},B=function(){D(),t.req&&t.req.abort()};e.magnificPopup.registerModule(O,{options:{settings:null,cursor:"mfp-ajax-cur",tError:'<a href="%url%">The content</a> could not be loaded.'},proto:{initAjax:function(){t.types.push(O),A=t.st.ajax.cursor,N(a+"."+O,B),N("BeforeChange."+O,B)},getAjax:function(n){A&&e(document.body).addClass(A),t.updateStatus("loading");var r=e.extend({url:n.src,success:function(r,o,i){var s={data:r,xhr:i};C("ParseAjax",s),t.appendContent(e(s.data),O),n.finished=!0,D(),t._setFocus(),setTimeout(function(){t.wrap.addClass(g)},16),t.updateStatus("ready"),C("AjaxContentAdded")},error:function(){D(),n.finished=n.loadError=!0,t.updateStatus("error",t.st.ajax.tError.replace("%url%",n.src))}},t.st.ajax.settings);return t.req=e.ajax(r),""}}});var L,P=function(n){if(n.data&&void 0!==n.data.title)return n.data.title;var r=t.st.image.titleSrc;if(r){if(e.isFunction(r))return r.call(t,n);if(n.el)return n.el.attr(r)||""}return""};e.magnificPopup.registerModule("image",{options:{markup:'<div class="mfp-figure"><div class="mfp-close"></div><figure><div class="mfp-img"></div><figcaption><div class="mfp-bottom-bar"><div class="mfp-title"></div><div class="mfp-counter"></div></div></figcaption></figure></div>',cursor:"mfp-zoom-out-cur",titleSrc:"title",verticalFit:!0,tError:'<a href="%url%">The image</a> could not be loaded.'},proto:{initImage:function(){var n=t.st.image,r=".image";t.types.push("image"),N(p+r,function(){"image"===t.currItem.type&&n.cursor&&e(document.body).addClass(n.cursor)}),N(a+r,function(){n.cursor&&e(document.body).removeClass(n.cursor),w.off("resize"+m)}),N("Resize"+r,t.resizeImage),t.isLowIE&&N("AfterChange",t.resizeImage)},resizeImage:function(){var e=t.currItem;if(e&&e.img&&t.st.image.verticalFit){var n=0;t.isLowIE&&(n=parseInt(e.img.css("padding-top"),10)+parseInt(e.img.css("padding-bottom"),10)),e.img.css("max-height",t.wH-n)}},_onImageHasSize:function(e){e.img&&(e.hasSize=!0,L&&clearInterval(L),e.isCheckingImgSize=!1,C("ImageHasSize",e),e.imgHidden&&(t.content&&t.content.removeClass("mfp-loading"),e.imgHidden=!1))},findImageSize:function(e){var n=0,r=e.img[0],o=function(i){L&&clearInterval(L),L=setInterval(function(){return r.naturalWidth>0?void t._onImageHasSize(e):(n>200&&clearInterval(L),n++,void(3===n?o(10):40===n?o(50):100===n&&o(500)))},i)};o(1)},getImage:function(n,r){var o=0,i=function(){n&&(n.img[0].complete?(n.img.off(".mfploader"),n===t.currItem&&(t._onImageHasSize(n),t.updateStatus("ready")),n.hasSize=!0,n.loaded=!0,C("ImageLoadComplete")):(o++,200>o?setTimeout(i,100):s()))},s=function(){n&&(n.img.off(".mfploader"),n===t.currItem&&(t._onImageHasSize(n),t.updateStatus("error",a.tError.replace("%url%",n.src))),n.hasSize=!0,n.loaded=!0,n.loadError=!0)},a=t.st.image,l=r.find(".mfp-img");if(l.length){var c=document.createElement("img");c.className="mfp-img",n.el&&n.el.find("img").length&&(c.alt=n.el.find("img").attr("alt")),n.img=e(c).on("load.mfploader",i).on("error.mfploader",s),c.src=n.src,l.is("img")&&(n.img=n.img.clone()),c=n.img[0],c.naturalWidth>0?n.hasSize=!0:c.width||(n.hasSize=!1)}return t._parseMarkup(r,{title:P(n),img_replaceWith:n.img},n),t.resizeImage(),n.hasSize?(L&&clearInterval(L),n.loadError?(r.addClass("mfp-loading"),t.updateStatus("error",a.tError.replace("%url%",n.src))):(r.removeClass("mfp-loading"),t.updateStatus("ready")),r):(t.updateStatus("loading"),n.loading=!0,n.hasSize||(n.imgHidden=!0,r.addClass("mfp-loading"),t.findImageSize(n)),r)}}});var q,z=function(){return void 0===q&&(q=void 0!==document.createElement("p").style.MozTransform),q};e.magnificPopup.registerModule("zoom",{options:{enabled:!1,easing:"ease-in-out",duration:300,opener:function(e){return e.is("img")?e:e.find("img")}},proto:{initZoom:function(){var e,n=t.st.zoom,r=".zoom";if(n.enabled&&t.supportsTransition){var o,i,s=n.duration,c=function(e){var t=e.clone().removeAttr("style").removeAttr("class").addClass("mfp-animated-image"),r="all "+n.duration/1e3+"s "+n.easing,o={position:"fixed",zIndex:9999,left:0,top:0,"-webkit-backface-visibility":"hidden"},i="transition";return o["-webkit-"+i]=o["-moz-"+i]=o["-o-"+i]=o[i]=r,t.css(o),t},u=function(){t.content.css("visibility","visible")};N("BuildControls"+r,function(){if(t._allowZoom()){if(clearTimeout(o),t.content.css("visibility","hidden"),e=t._getItemToZoom(),!e)return void u();i=c(e),i.css(t._getOffset()),t.wrap.append(i),o=setTimeout(function(){i.css(t._getOffset(!0)),o=setTimeout(function(){u(),setTimeout(function(){i.remove(),e=i=null,C("ZoomAnimationEnded")},16)},s)},16)}}),N(l+r,function(){if(t._allowZoom()){if(clearTimeout(o),t.st.removalDelay=s,!e){if(e=t._getItemToZoom(),!e)return;i=c(e)}i.css(t._getOffset(!0)),t.wrap.append(i),t.content.css("visibility","hidden"),setTimeout(function(){i.css(t._getOffset())},16)}}),N(a+r,function(){t._allowZoom()&&(u(),i&&i.remove(),e=null)})}},_allowZoom:function(){return"image"===t.currItem.type},_getItemToZoom:function(){return t.currItem.hasSize?t.currItem.img:!1},_getOffset:function(n){var r;r=n?t.currItem.img:t.st.zoom.opener(t.currItem.el||t.currItem);var o=r.offset(),i=parseInt(r.css("padding-top"),10),s=parseInt(r.css("padding-bottom"),10);o.top-=e(window).scrollTop()-i;var a={width:r.width(),height:(b?r.innerHeight():r[0].offsetHeight)-s-i};return z()?a["-moz-transform"]=a.transform="translate("+o.left+"px,"+o.top+"px)":(a.left=o.left,a.top=o.top),a}}});var $="iframe",H="//about:blank",R=function(e){if(t.currTemplate[$]){var n=t.currTemplate[$].find("iframe");n.length&&(e||(n[0].src=H),t.isIE8&&n.css("display",e?"block":"none"))}};e.magnificPopup.registerModule($,{options:{markup:'<div class="mfp-iframe-scaler"><div class="mfp-close"></div><iframe class="mfp-iframe" src="//about:blank" frameborder="0" allowfullscreen></iframe></div>',srcAction:"iframe_src",patterns:{youtube:{index:"youtube.com",id:"v=",src:"//www.youtube.com/embed/%id%?autoplay=1"},vimeo:{index:"vimeo.com/",id:"/",src:"//player.vimeo.com/video/%id%?autoplay=1"},gmaps:{index:"//maps.google.",src:"%id%&output=embed"}}},proto:{initIframe:function(){t.types.push($),N("BeforeChange",function(e,t,n){t!==n&&(t===$?R():n===$&&R(!0))}),N(a+"."+$,function(){R()})},getIframe:function(n,r){var o=n.src,i=t.st.iframe;e.each(i.patterns,function(){return o.indexOf(this.index)>-1?(this.id&&(o="string"==typeof this.id?o.substr(o.lastIndexOf(this.id)+this.id.length,o.length):this.id.call(this,o)),o=this.src.replace("%id%",o),!1):void 0});var s={};return i.srcAction&&(s[i.srcAction]=o),t._parseMarkup(r,s,n),t.updateStatus("ready"),r}}});var U=function(e){var n=t.items.length;return e>n-1?e-n:0>e?n+e:e},W=function(e,t,n){return e.replace(/%curr%/gi,t+1).replace(/%total%/gi,n)};e.magnificPopup.registerModule("gallery",{options:{enabled:!1,arrowMarkup:'<button title="%title%" type="button" class="mfp-arrow mfp-arrow-%dir%"></button>',preload:[0,2],navigateByImgClick:!0,arrows:!0,tPrev:"Previous (Left arrow key)",tNext:"Next (Right arrow key)",tCounter:"%curr% of %total%"},proto:{initGallery:function(){var n=t.st.gallery,o=".mfp-gallery",s=Boolean(e.fn.mfpFastClick);return t.direction=!0,n&&n.enabled?(i+=" mfp-gallery",N(p+o,function(){n.navigateByImgClick&&t.wrap.on("click"+o,".mfp-img",function(){return t.items.length>1?(t.next(),!1):void 0}),r.on("keydown"+o,function(e){37===e.keyCode?t.prev():39===e.keyCode&&t.next()})}),N("UpdateStatus"+o,function(e,n){n.text&&(n.text=W(n.text,t.currItem.index,t.items.length))}),N(d+o,function(e,r,o,i){var s=t.items.length;o.counter=s>1?W(n.tCounter,i.index,s):""}),N("BuildControls"+o,function(){if(t.items.length>1&&n.arrows&&!t.arrowLeft){var r=n.arrowMarkup,o=t.arrowLeft=e(r.replace(/%title%/gi,n.tPrev).replace(/%dir%/gi,"left")).addClass(v),i=t.arrowRight=e(r.replace(/%title%/gi,n.tNext).replace(/%dir%/gi,"right")).addClass(v),a=s?"mfpFastClick":"click";o[a](function(){t.prev()}),i[a](function(){t.next()}),t.isIE7&&(x("b",o[0],!1,!0),x("a",o[0],!1,!0),x("b",i[0],!1,!0),x("a",i[0],!1,!0)),t.container.append(o.add(i))}}),N(h+o,function(){t._preloadTimeout&&clearTimeout(t._preloadTimeout),t._preloadTimeout=setTimeout(function(){t.preloadNearbyImages(),t._preloadTimeout=null},16)}),void N(a+o,function(){r.off(o),t.wrap.off("click"+o),t.arrowLeft&&s&&t.arrowLeft.add(t.arrowRight).destroyMfpFastClick(),t.arrowRight=t.arrowLeft=null})):!1},next:function(){t.direction=!0,t.index=U(t.index+1),t.updateItemHTML()},prev:function(){t.direction=!1,t.index=U(t.index-1),t.updateItemHTML()},goTo:function(e){t.direction=e>=t.index,t.index=e,t.updateItemHTML()},preloadNearbyImages:function(){var e,n=t.st.gallery.preload,r=Math.min(n[0],t.items.length),o=Math.min(n[1],t.items.length);for(e=1;e<=(t.direction?o:r);e++)t._preloadItem(t.index+e);for(e=1;e<=(t.direction?r:o);e++)t._preloadItem(t.index-e)},_preloadItem:function(n){if(n=U(n),!t.items[n].preloaded){var r=t.items[n];r.parsed||(r=t.parseEl(n)),C("LazyLoad",r),"image"===r.type&&(r.img=e('<img class="mfp-img" />').on("load.mfploader",function(){r.hasSize=!0}).on("error.mfploader",function(){r.hasSize=!0,r.loadError=!0,C("LazyLoadError",r)}).attr("src",r.src)),r.preloaded=!0}}}});var Z="retina";e.magnificPopup.registerModule(Z,{options:{replaceSrc:function(e){return e.src.replace(/\.\w+$/,function(e){return"@2x"+e})},ratio:1},proto:{initRetina:function(){if(window.devicePixelRatio>1){var e=t.st.retina,n=e.ratio;n=isNaN(n)?n():n,n>1&&(N("ImageHasSize."+Z,function(e,t){t.img.css({"max-width":t.img[0].naturalWidth/n,width:"100%"})}),N("ElementParse."+Z,function(t,r){r.src=e.replaceSrc(r,n)}))}}}}),function(){var t=1e3,n="ontouchstart"in window,r=function(){w.off("touchmove"+i+" touchend"+i)},o="mfpFastClick",i="."+o;e.fn.mfpFastClick=function(o){return e(this).each(function(){var s,a=e(this);if(n){var l,c,u,d,p,h;a.on("touchstart"+i,function(e){d=!1,h=1,p=e.originalEvent?e.originalEvent.touches[0]:e.touches[0],c=p.clientX,u=p.clientY,w.on("touchmove"+i,function(e){p=e.originalEvent?e.originalEvent.touches:e.touches,h=p.length,p=p[0],(Math.abs(p.clientX-c)>10||Math.abs(p.clientY-u)>10)&&(d=!0,r())}).on("touchend"+i,function(e){r(),d||h>1||(s=!0,e.preventDefault(),clearTimeout(l),l=setTimeout(function(){s=!1},t),o())})})}a.on("click"+i,function(){s||o()})})},e.fn.destroyMfpFastClick=function(){e(this).off("touchstart"+i+" click"+i),n&&w.off("touchmove"+i+" touchend"+i)}}(),k()}),function(e){if("function"==typeof define&&define.amd)define(e);else if("object"==typeof exports)module.exports=e();else{var t=window.Cookies,n=window.Cookies=e();n.noConflict=function(){return window.Cookies=t,n}}}(function(){function e(){for(var e=0,t={};e<arguments.length;e++){var n=arguments[e];for(var r in n)t[r]=n[r]}return t}function t(n){function r(t,o,i){var s;if(arguments.length>1){if(i=e({path:"/"},r.defaults,i),"number"==typeof i.expires){var a=new Date;a.setMilliseconds(a.getMilliseconds()+864e5*i.expires),i.expires=a}try{s=JSON.stringify(o),/^[\{\[]/.test(s)&&(o=s)}catch(l){}return o=n.write?n.write(o,t):encodeURIComponent(String(o)).replace(/%(23|24|26|2B|3A|3C|3E|3D|2F|3F|40|5B|5D|5E|60|7B|7D|7C)/g,decodeURIComponent),t=encodeURIComponent(String(t)),t=t.replace(/%(23|24|26|2B|5E|60|7C)/g,decodeURIComponent),t=t.replace(/[\(\)]/g,escape),document.cookie=[t,"=",o,i.expires&&"; expires="+i.expires.toUTCString(),i.path&&"; path="+i.path,i.domain&&"; domain="+i.domain,i.secure?"; secure":""].join("")}t||(s={});for(var c=document.cookie?document.cookie.split("; "):[],u=/(%[0-9A-Z]{2})+/g,d=0;d<c.length;d++){var p=c[d].split("="),h=p[0].replace(u,decodeURIComponent),f=p.slice(1).join("=");'"'===f.charAt(0)&&(f=f.slice(1,-1));try{if(f=n.read?n.read(f,h):n(f,h)||f.replace(u,decodeURIComponent),this.json)try{f=JSON.parse(f)}catch(l){}if(t===h){s=f;break}t||(s[h]=f)}catch(l){}}return s}return r.get=r.set=r,r.getJSON=function(){return r.apply({json:!0},[].slice.call(arguments))},r.defaults={},r.remove=function(t,n){r(t,"",e(n,{expires:-1}))},r.withConverter=t,r}return t(function(){})}),!function(e,t,n){function r(e,t){return typeof e===t}function o(){var e,t,n,o,i,l,u;for(var d in a)if(a.hasOwnProperty(d)){if(e=[],t=a[d],t.name&&(e.push(t.name.toLowerCase()),t.options&&t.options.aliases&&t.options.aliases.length))for(n=0;n<t.options.aliases.length;n++)e.push(t.options.aliases[n].toLowerCase());for(o=r(t.fn,"function")?t.fn():t.fn,i=0;i<e.length;i++)l=e[i],u=l.split("."),1===u.length?c[u[0]]=o:(!c[u[0]]||c[u[0]]instanceof Boolean||(c[u[0]]=new Boolean(c[u[0]])),c[u[0]][u[1]]=o),s.push((o?"":"no-")+u.join("-"))}}function i(e){var t=u.className,n=c._config.classPrefix||"";if(d&&(t=t.baseVal),c._config.enableJSClass){var r=new RegExp("(^|\\s)"+n+"no-js(\\s|$)");t=t.replace(r,"$1"+n+"js$2")}c._config.enableClasses&&(t+=" "+n+e.join(" "+n),d?u.className.baseVal=t:u.className=t)}var s=[],a=[],l={_version:"3.1.0",_config:{classPrefix:"",enableClasses:!0,enableJSClass:!0,usePrefixes:!0},_q:[],on:function(e,t){var n=this;setTimeout(function(){t(n[e])},0)},addTest:function(e,t,n){a.push({name:e,fn:t,options:n})},addAsyncTest:function(e){a.push({name:null,fn:e})}},c=function(){};c.prototype=l,c=new c;var u=t.documentElement,d="svg"===u.nodeName.toLowerCase();o(),i(s),delete l.addTest,delete l.addAsyncTest;for(var p=0;p<c._q.length;p++)c._q[p]();e.Modernizr=c}(window,document),function e(t,n,r){function o(s,a){if(!n[s]){if(!t[s]){var l="function"==typeof require&&require;if(!a&&l)return l(s,!0);if(i)return i(s,!0);var c=new Error("Cannot find module '"+s+"'");throw c.code="MODULE_NOT_FOUND",c}var u=n[s]={exports:{}};t[s][0].call(u.exports,function(e){var n=t[s][1][e];return o(n?n:e)},u,u.exports,e,t,n,r)}return n[s].exports}for(var i="function"==typeof require&&require,s=0;s<r.length;s++)o(r[s]);return o}({1:[function(e,t,n){var r,o,i,s,a,l,c,u=function(e,t){function n(){this.constructor=e}for(var r in t)d.call(t,r)&&(e[r]=t[r]);return n.prototype=t.prototype,e.prototype=new n,e.__super__=t.prototype,e},d={}.hasOwnProperty;c=e("./node"),a=c.Position,r=jQuery,o=function(){function e(e){this.tree_widget=e,this.hovered_area=null,this.$ghost=null,this.hit_areas=[],this.is_dragging=!1,this.current_item=null}return e.prototype.mouseCapture=function(e){var t,n;return t=r(e.target),this.mustCaptureElement(t)?this.tree_widget.options.onIsMoveHandle&&!this.tree_widget.options.onIsMoveHandle(t)?null:(n=this.tree_widget._getNodeElement(t),n&&this.tree_widget.options.onCanMove&&(this.tree_widget.options.onCanMove(n.node)||(n=null)),this.current_item=n,null!==this.current_item):null},e.prototype.mouseStart=function(e){var t;return this.refresh(),t=r(e.target).offset(),this.drag_element=new i(this.current_item.node,e.page_x-t.left,e.page_y-t.top,this.tree_widget.element),this.is_dragging=!0,this.current_item.$element.addClass("jqtree-moving"),!0},e.prototype.mouseDrag=function(e){var t,n;return this.drag_element.move(e.page_x,e.page_y),t=this.findHoveredArea(e.page_x,e.page_y),n=this.canMoveToArea(t),n&&t?(t.node.isFolder()||this.stopOpenFolderTimer(),this.hovered_area!==t&&(this.hovered_area=t,this.mustOpenFolderTimer(t)?this.startOpenFolderTimer(t.node):this.stopOpenFolderTimer(),this.updateDropHint())):(this.removeHover(),this.removeDropHint(),this.stopOpenFolderTimer()),t||null!=this.tree_widget.options.onDragMove&&this.tree_widget.options.onDragMove(this.current_item.node,e.original_event),!0},e.prototype.mustCaptureElement=function(e){return!e.is("input,select")},e.prototype.canMoveToArea=function(e){var t;return e?this.tree_widget.options.onCanMoveTo?(t=a.getName(e.position),this.tree_widget.options.onCanMoveTo(this.current_item.node,e.node,t)):!0:!1},e.prototype.mouseStop=function(e){var t;return this.moveItem(e),this.clear(),this.removeHover(),this.removeDropHint(),this.removeHitAreas(),t=this.current_item,this.current_item&&(this.current_item.$element.removeClass("jqtree-moving"),this.current_item=null),this.is_dragging=!1,!this.hovered_area&&t&&null!=this.tree_widget.options.onDragStop&&this.tree_widget.options.onDragStop(t.node,e.original_event),!1},e.prototype.refresh=function(){return this.removeHitAreas(),this.current_item&&(this.generateHitAreas(),this.current_item=this.tree_widget._getNodeElementForNode(this.current_item.node),this.is_dragging)?this.current_item.$element.addClass("jqtree-moving"):void 0},e.prototype.removeHitAreas=function(){return this.hit_areas=[]},e.prototype.clear=function(){return this.drag_element.remove(),this.drag_element=null},e.prototype.removeDropHint=function(){return this.previous_ghost?this.previous_ghost.remove():void 0},e.prototype.removeHover=function(){return this.hovered_area=null},e.prototype.generateHitAreas=function(){var e;return e=new s(this.tree_widget.tree,this.current_item.node,this.getTreeDimensions().bottom),this.hit_areas=e.generate()},e.prototype.findHoveredArea=function(e,t){var n,r,o,i,s;if(r=this.getTreeDimensions(),e<r.left||t<r.top||e>r.right||t>r.bottom)return null;for(i=0,o=this.hit_areas.length;o>i;)if(s=i+o>>1,n=this.hit_areas[s],t<n.top)o=s;else{if(!(t>n.bottom))return n;i=s+1}return null},e.prototype.mustOpenFolderTimer=function(e){var t;return t=e.node,t.isFolder()&&!t.is_open&&e.position===a.INSIDE},e.prototype.updateDropHint=function(){var e;if(this.hovered_area)return this.removeDropHint(),e=this.tree_widget._getNodeElementForNode(this.hovered_area.node),this.previous_ghost=e.addDropHint(this.hovered_area.position)},e.prototype.startOpenFolderTimer=function(e){var t;return t=function(t){return function(){return t.tree_widget._openNode(e,t.tree_widget.options.slide,function(){return t.refresh(),t.updateDropHint()})}}(this),this.stopOpenFolderTimer(),this.open_folder_timer=setTimeout(t,this.tree_widget.options.openFolderDelay)},e.prototype.stopOpenFolderTimer=function(){return this.open_folder_timer?(clearTimeout(this.open_folder_timer),this.open_folder_timer=null):void 0},e.prototype.moveItem=function(e){var t,n,r,o,i,s;return this.hovered_area&&this.hovered_area.position!==a.NONE&&this.canMoveToArea(this.hovered_area)&&(r=this.current_item.node,s=this.hovered_area.node,o=this.hovered_area.position,i=r.parent,o===a.INSIDE&&(this.hovered_area.node.is_open=!0),t=function(e){return function(){return e.tree_widget.tree.moveNode(r,s,o),e.tree_widget.element.empty(),e.tree_widget._refreshElements()}}(this),n=this.tree_widget._triggerEvent("tree.move",{move_info:{moved_node:r,target_node:s,position:a.getName(o),previous_parent:i,do_move:t,original_event:e.original_event}}),!n.isDefaultPrevented())?t():void 0},e.prototype.getTreeDimensions=function(){var e;return e=this.tree_widget.element.offset(),{left:e.left,top:e.top,right:e.left+this.tree_widget.element.width(),bottom:e.top+this.tree_widget.element.height()+16}},e}(),l=function(){function e(e){this.tree=e}return e.prototype.iterate=function(){var e,t;return t=!0,(e=function(n){return function(o,i){var s,a,l,c,u,d,p,h;if(p=(o.is_open||!o.element)&&o.hasChildren(),o.element){if(s=r(o.element),!s.is(":visible"))return;t&&(n.handleFirstNode(o,s),t=!1),o.hasChildren()?o.is_open?n.handleOpenFolder(o,s)||(p=!1):n.handleClosedFolder(o,i,s):n.handleNode(o,i,s)}if(p){for(l=o.children.length,h=o.children,c=u=0,d=h.length;d>u;c=++u)a=h[c],c===l-1?e(o.children[c],null):e(o.children[c],o.children[c+1]);if(o.is_open)return n.handleAfterOpenFolder(o,i,s)}}}(this))(this.tree,null)},e.prototype.handleNode=function(e,t,n){},e.prototype.handleOpenFolder=function(e,t){},e.prototype.handleClosedFolder=function(e,t,n){},e.prototype.handleAfterOpenFolder=function(e,t,n){},e.prototype.handleFirstNode=function(e,t){},e}(),s=function(e){function t(e,n,r){t.__super__.constructor.call(this,e),this.current_node=n,this.tree_bottom=r}return u(t,e),t.prototype.generate=function(){return this.positions=[],this.last_top=0,this.iterate(),this.generateHitAreas(this.positions)},t.prototype.getTop=function(e){return e.offset().top},t.prototype.addPosition=function(e,t,n){var r;return r={top:n,node:e,position:t},this.positions.push(r),this.last_top=n},t.prototype.handleNode=function(e,t,n){var r;return r=this.getTop(n),e===this.current_node?this.addPosition(e,a.NONE,r):this.addPosition(e,a.INSIDE,r),
+t===this.current_node||e===this.current_node?this.addPosition(e,a.NONE,r):this.addPosition(e,a.AFTER,r)},t.prototype.handleOpenFolder=function(e,t){return e===this.current_node?!1:(e.children[0]!==this.current_node&&this.addPosition(e,a.INSIDE,this.getTop(t)),!0)},t.prototype.handleClosedFolder=function(e,t,n){var r;return r=this.getTop(n),e===this.current_node?this.addPosition(e,a.NONE,r):(this.addPosition(e,a.INSIDE,r),t!==this.current_node?this.addPosition(e,a.AFTER,r):void 0)},t.prototype.handleFirstNode=function(e,t){return e!==this.current_node?this.addPosition(e,a.BEFORE,this.getTop(r(e.element))):void 0},t.prototype.handleAfterOpenFolder=function(e,t,n){return e===this.current_node.node||t===this.current_node.node?this.addPosition(e,a.NONE,this.last_top):this.addPosition(e,a.AFTER,this.last_top)},t.prototype.generateHitAreas=function(e){var t,n,r,o,i,s;for(s=-1,t=[],n=[],r=0,o=e.length;o>r;r++)i=e[r],i.top!==s&&t.length&&(t.length&&this.generateHitAreasForGroup(n,t,s,i.top),s=i.top,t=[]),t.push(i);return this.generateHitAreasForGroup(n,t,s,this.tree_bottom),n},t.prototype.generateHitAreasForGroup=function(e,t,n,r){var o,i,s,a,l;for(l=Math.min(t.length,4),o=Math.round((r-n)/l),i=n,s=0;l>s;)a=t[s],e.push({top:i,bottom:i+o,node:a.node,position:a.position}),i+=o,s+=1;return null},t}(l),i=function(){function e(e,t,n,o){this.offset_x=t,this.offset_y=n,this.$element=r('<span class="jqtree-title jqtree-dragging">'+e.name+"</span>"),this.$element.css("position","absolute"),o.append(this.$element)}return e.prototype.move=function(e,t){return this.$element.offset({left:e-this.offset_x,top:t-this.offset_y})},e.prototype.remove=function(){return this.$element.remove()},e}(),t.exports={DragAndDropHandler:o,DragElement:i,HitAreasGenerator:s}},{"./node":5}],2:[function(e,t,n){var r,o,i,s,a,l;a=e("./node_element"),i=a.NodeElement,l=e("./util"),s=l.html_escape,r=jQuery,o=function(){function e(e){this.tree_widget=e,this.opened_icon_element=this.createButtonElement(e.options.openedIcon),this.closed_icon_element=this.createButtonElement(e.options.closedIcon)}return e.prototype.render=function(e){return e&&e.parent?this.renderFromNode(e):this.renderFromRoot()},e.prototype.renderFromRoot=function(){var e;return e=this.tree_widget.element,e.empty(),this.createDomElements(e[0],this.tree_widget.tree.children,!0,!0,1)},e.prototype.renderFromNode=function(e){var t,n;return t=r(e.element),n=this.createLi(e),this.attachNodeData(e,n),t.after(n),t.remove(),e.children?this.createDomElements(n,e.children,!1,!1,e.getLevel()):void 0},e.prototype.createDomElements=function(e,t,n,r,o){var i,s,a,l,c;for(c=this.createUl(n),e.appendChild(c),s=0,a=t.length;a>s;s++)i=t[s],l=this.createLi(i,o),c.appendChild(l),this.attachNodeData(i,l),i.hasChildren()&&this.createDomElements(l,i.children,!1,i.is_open,o+1);return null},e.prototype.attachNodeData=function(e,t){return e.element=t,r(t).data("node",e)},e.prototype.createUl=function(e){var t,n,r;return e?(t="jqtree-tree",n="tree",this.tree_widget.options.rtl&&(t+=" jqtree-rtl")):(t="",n="group"),r=document.createElement("ul"),r.className="jqtree_common "+t,r.setAttribute("role",n),r},e.prototype.createLi=function(e,t){var n,o;return n=this.tree_widget.select_node_handler&&this.tree_widget.select_node_handler.isNodeSelected(e),o=e.isFolder()?this.createFolderLi(e,t,n):this.createNodeLi(e,t,n),this.tree_widget.options.onCreateLi&&this.tree_widget.options.onCreateLi(e,r(o)),o},e.prototype.createFolderLi=function(e,t,n){var r,o,i,s,a,l,c;return r=this.getButtonClasses(e),s=this.getFolderClasses(e,n),a=e.is_open?this.opened_icon_element:this.closed_icon_element,c=document.createElement("li"),c.className="jqtree_common "+s,c.setAttribute("role","presentation"),i=document.createElement("div"),i.className="jqtree-element jqtree_common",i.setAttribute("role","presentation"),c.appendChild(i),o=document.createElement("a"),o.className=r,o.appendChild(a.cloneNode(!1)),o.setAttribute("role","presentation"),o.setAttribute("aria-hidden","true"),this.tree_widget.options.buttonLeft&&i.appendChild(o),i.appendChild(this.createTitleSpan(e.name,t,n,e.is_open,l=!0)),this.tree_widget.options.buttonLeft||i.appendChild(o),c},e.prototype.createNodeLi=function(e,t,n){var r,o,i,s,a;return a=["jqtree_common"],n&&a.push("jqtree-selected"),r=a.join(" "),s=document.createElement("li"),s.className=r,s.setAttribute("role","presentation"),o=document.createElement("div"),o.className="jqtree-element jqtree_common",o.setAttribute("role","presentation"),s.appendChild(o),o.appendChild(this.createTitleSpan(e.name,t,n,e.is_open,i=!1)),s},e.prototype.createTitleSpan=function(e,t,n,r,o){var i,s;return s=document.createElement("span"),i="jqtree-title jqtree_common",o&&(i+=" jqtree-title-folder"),s.className=i,s.setAttribute("role","treeitem"),s.setAttribute("aria-level",t),s.setAttribute("aria-selected",l.getBoolString(n)),s.setAttribute("aria-expanded",l.getBoolString(r)),n&&s.setAttribute("tabindex",0),s.innerHTML=this.escapeIfNecessary(e),s},e.prototype.getButtonClasses=function(e){var t;return t=["jqtree-toggler","jqtree_common"],e.is_open||t.push("jqtree-closed"),this.tree_widget.options.buttonLeft?t.push("jqtree-toggler-left"):t.push("jqtree-toggler-right"),t.join(" ")},e.prototype.getFolderClasses=function(e,t){var n;return n=["jqtree-folder"],e.is_open||n.push("jqtree-closed"),t&&n.push("jqtree-selected"),e.is_loading&&n.push("jqtree-loading"),n.join(" ")},e.prototype.escapeIfNecessary=function(e){return this.tree_widget.options.autoEscape?s(e):e},e.prototype.createButtonElement=function(e){var t;return"string"==typeof e?(t=document.createElement("div"),t.innerHTML=e,document.createTextNode(t.innerHTML)):r(e)[0]},e}(),t.exports=o},{"./node_element":6,"./util":12}],3:[function(e,t,n){var r,o,i=function(e,t){return function(){return e.apply(t,arguments)}};r=jQuery,o=function(){function e(e){this.selectNode=i(this.selectNode,this),this.tree_widget=e,e.options.keyboardSupport&&r(document).bind("keydown.jqtree",r.proxy(this.handleKeyDown,this))}var t,n,o,s;return n=37,s=38,o=39,t=40,e.prototype.deinit=function(){return r(document).unbind("keydown.jqtree")},e.prototype.moveDown=function(){var e;return e=this.tree_widget.getSelectedNode(),e?this.selectNode(e.getNextNode()):!1},e.prototype.moveUp=function(){var e;return e=this.tree_widget.getSelectedNode(),e?this.selectNode(e.getPreviousNode()):!1},e.prototype.moveRight=function(){var e;return e=this.tree_widget.getSelectedNode(),e&&e.isFolder()?e.is_open?this.selectNode(e.getNextNode()):(this.tree_widget.openNode(e),!1):!0},e.prototype.moveLeft=function(){var e;return e=this.tree_widget.getSelectedNode(),e?e.isFolder()&&e.is_open?(this.tree_widget.closeNode(e),!1):this.selectNode(e.getParent()):!0},e.prototype.handleKeyDown=function(e){var i;if(!this.tree_widget.options.keyboardSupport)return!0;if(r(document.activeElement).is("textarea,input,select"))return!0;if(!this.tree_widget.getSelectedNode())return!0;switch(i=e.which){case t:return this.moveDown();case s:return this.moveUp();case o:return this.moveRight();case n:return this.moveLeft()}return!0},e.prototype.selectNode=function(e){return e?(this.tree_widget.selectNode(e),this.tree_widget.scroll_handler&&!this.tree_widget.scroll_handler.isScrolledIntoView(r(e.element).find(".jqtree-element"))&&this.tree_widget.scrollToNode(e),!1):!0},e}(),t.exports=o},{}],4:[function(e,t,n){var r,o,i,s=function(e,t){function n(){this.constructor=e}for(var r in t)a.call(t,r)&&(e[r]=t[r]);return n.prototype=t.prototype,e.prototype=new n,e.__super__=t.prototype,e},a={}.hasOwnProperty;i=e("./simple.widget"),r=jQuery,o=function(e){function t(){return t.__super__.constructor.apply(this,arguments)}return s(t,e),t.is_mouse_handled=!1,t.prototype._init=function(){return this.$el.bind("mousedown.mousewidget",r.proxy(this._mouseDown,this)),this.$el.bind("touchstart.mousewidget",r.proxy(this._touchStart,this)),this.is_mouse_started=!1,this.mouse_delay=0,this._mouse_delay_timer=null,this._is_mouse_delay_met=!0,this.mouse_down_info=null},t.prototype._deinit=function(){var e;return this.$el.unbind("mousedown.mousewidget"),this.$el.unbind("touchstart.mousewidget"),e=r(document),e.unbind("mousemove.mousewidget"),e.unbind("mouseup.mousewidget")},t.prototype._mouseDown=function(e){var t;if(1===e.which)return t=this._handleMouseDown(e,this._getPositionInfo(e)),t&&e.preventDefault(),t},t.prototype._handleMouseDown=function(e,n){return!t.is_mouse_handled&&(this.is_mouse_started&&this._handleMouseUp(n),this.mouse_down_info=n,this._mouseCapture(n))?(this._handleStartMouse(),this.is_mouse_handled=!0,!0):void 0},t.prototype._handleStartMouse=function(){var e;return e=r(document),e.bind("mousemove.mousewidget",r.proxy(this._mouseMove,this)),e.bind("touchmove.mousewidget",r.proxy(this._touchMove,this)),e.bind("mouseup.mousewidget",r.proxy(this._mouseUp,this)),e.bind("touchend.mousewidget",r.proxy(this._touchEnd,this)),this.mouse_delay?this._startMouseDelayTimer():void 0},t.prototype._startMouseDelayTimer=function(){return this._mouse_delay_timer&&clearTimeout(this._mouse_delay_timer),this._mouse_delay_timer=setTimeout(function(e){return function(){return e._is_mouse_delay_met=!0}}(this),this.mouse_delay),this._is_mouse_delay_met=!1},t.prototype._mouseMove=function(e){return this._handleMouseMove(e,this._getPositionInfo(e))},t.prototype._handleMouseMove=function(e,t){return this.is_mouse_started?(this._mouseDrag(t),e.preventDefault()):this.mouse_delay&&!this._is_mouse_delay_met?!0:(this.is_mouse_started=this._mouseStart(this.mouse_down_info)!==!1,this.is_mouse_started?this._mouseDrag(t):this._handleMouseUp(t),!this.is_mouse_started)},t.prototype._getPositionInfo=function(e){return{page_x:e.pageX,page_y:e.pageY,target:e.target,original_event:e}},t.prototype._mouseUp=function(e){return this._handleMouseUp(this._getPositionInfo(e))},t.prototype._handleMouseUp=function(e){var t;t=r(document),t.unbind("mousemove.mousewidget"),t.unbind("touchmove.mousewidget"),t.unbind("mouseup.mousewidget"),t.unbind("touchend.mousewidget"),this.is_mouse_started&&(this.is_mouse_started=!1,this._mouseStop(e))},t.prototype._mouseCapture=function(e){return!0},t.prototype._mouseStart=function(e){return null},t.prototype._mouseDrag=function(e){return null},t.prototype._mouseStop=function(e){return null},t.prototype.setMouseDelay=function(e){return this.mouse_delay=e},t.prototype._touchStart=function(e){var t;if(!(e.originalEvent.touches.length>1))return t=e.originalEvent.changedTouches[0],this._handleMouseDown(e,this._getPositionInfo(t))},t.prototype._touchMove=function(e){var t;if(!(e.originalEvent.touches.length>1))return t=e.originalEvent.changedTouches[0],this._handleMouseMove(e,this._getPositionInfo(t))},t.prototype._touchEnd=function(e){var t;if(!(e.originalEvent.touches.length>1))return t=e.originalEvent.changedTouches[0],this._handleMouseUp(this._getPositionInfo(t))},t}(i),t.exports=o},{"./simple.widget":10}],5:[function(e,t,n){var r,o,i;r=jQuery,i={getName:function(e){return i.strings[e-1]},nameToIndex:function(e){var t,n,r;for(t=n=1,r=i.strings.length;r>=1?r>=n:n>=r;t=r>=1?++n:--n)if(i.strings[t-1]===e)return t;return 0}},i.BEFORE=1,i.AFTER=2,i.INSIDE=3,i.NONE=4,i.strings=["before","after","inside","none"],o=function(){function e(t,n,r){null==n&&(n=!1),null==r&&(r=e),this.name="",this.setData(t),this.children=[],this.parent=null,n&&(this.id_mapping={},this.tree=this,this.node_class=r)}return e.prototype.setData=function(e){var t,n,r;if(n=function(e){return function(t){return null!==t?e.name=t:void 0}}(this),"object"!=typeof e)n(e);else for(t in e)r=e[t],"label"===t?n(r):this[t]=r;return null},e.prototype.initFromData=function(e){var t,n;return n=function(e){return function(n){return e.setData(n),n.children?t(n.children):void 0}}(this),t=function(e){return function(t){var n,r,o,i;for(r=0,o=t.length;o>r;r++)n=t[r],i=new e.tree.node_class(""),i.initFromData(n),e.addChild(i);return null}}(this),n(e),null},e.prototype.loadFromData=function(e){var t,n,r,o;for(this.removeChildren(),t=0,n=e.length;n>t;t++)o=e[t],r=new this.tree.node_class(o),this.addChild(r),"object"==typeof o&&o.children&&r.loadFromData(o.children);return null},e.prototype.addChild=function(e){return this.children.push(e),e._setParent(this)},e.prototype.addChildAtPosition=function(e,t){return this.children.splice(t,0,e),e._setParent(this)},e.prototype._setParent=function(e){return this.parent=e,this.tree=e.tree,this.tree.addNodeToIndex(this)},e.prototype.removeChild=function(e){return e.removeChildren(),this._removeChild(e)},e.prototype._removeChild=function(e){return this.children.splice(this.getChildIndex(e),1),this.tree.removeNodeFromIndex(e)},e.prototype.getChildIndex=function(e){return r.inArray(e,this.children)},e.prototype.hasChildren=function(){return 0!==this.children.length},e.prototype.isFolder=function(){return this.hasChildren()||this.load_on_demand},e.prototype.iterate=function(e){var t;return t=function(n,r){var o,i,s,a,l;if(n.children){for(a=n.children,i=0,s=a.length;s>i;i++)o=a[i],l=e(o,r),l&&o.hasChildren()&&t(o,r+1);return null}},t(this,0),null},e.prototype.moveNode=function(e,t,n){return e.isParentOf(t)?void 0:(e.parent._removeChild(e),n===i.AFTER?t.parent.addChildAtPosition(e,t.parent.getChildIndex(t)+1):n===i.BEFORE?t.parent.addChildAtPosition(e,t.parent.getChildIndex(t)):n===i.INSIDE?t.addChildAtPosition(e,0):void 0)},e.prototype.getData=function(e){var t;return null==e&&(e=!1),(t=function(e){var n,r,o,i,s,a,l;for(n=[],r=0,i=e.length;i>r;r++){s=e[r],a={};for(o in s)l=s[o],"parent"!==o&&"children"!==o&&"element"!==o&&"tree"!==o&&Object.prototype.hasOwnProperty.call(s,o)&&(a[o]=l);s.hasChildren()&&(a.children=t(s.children)),n.push(a)}return n})(e?[this]:this.children)},e.prototype.getNodeByName=function(e){var t;return t=null,this.iterate(function(n){return n.name===e?(t=n,!1):!0}),t},e.prototype.addAfter=function(e){var t,n;return this.parent?(n=new this.tree.node_class(e),t=this.parent.getChildIndex(this),this.parent.addChildAtPosition(n,t+1),n):null},e.prototype.addBefore=function(e){var t,n;return this.parent?(n=new this.tree.node_class(e),t=this.parent.getChildIndex(this),this.parent.addChildAtPosition(n,t),n):null},e.prototype.addParent=function(e){var t,n,r,o,i,s;if(this.parent){for(o=new this.tree.node_class(e),o._setParent(this.tree),i=this.parent,s=i.children,n=0,r=s.length;r>n;n++)t=s[n],o.addChild(t);return i.children=[],i.addChild(o),o}return null},e.prototype.remove=function(){return this.parent?(this.parent.removeChild(this),this.parent=null):void 0},e.prototype.append=function(e){var t;return t=new this.tree.node_class(e),this.addChild(t),t},e.prototype.prepend=function(e){var t;return t=new this.tree.node_class(e),this.addChildAtPosition(t,0),t},e.prototype.isParentOf=function(e){var t;for(t=e.parent;t;){if(t===this)return!0;t=t.parent}return!1},e.prototype.getLevel=function(){var e,t;for(e=0,t=this;t.parent;)e+=1,t=t.parent;return e},e.prototype.getNodeById=function(e){return this.id_mapping[e]},e.prototype.addNodeToIndex=function(e){return null!=e.id?this.id_mapping[e.id]=e:void 0},e.prototype.removeNodeFromIndex=function(e){return null!=e.id?delete this.id_mapping[e.id]:void 0},e.prototype.removeChildren=function(){return this.iterate(function(e){return function(t){return e.tree.removeNodeFromIndex(t),!0}}(this)),this.children=[]},e.prototype.getPreviousSibling=function(){var e;return this.parent?(e=this.parent.getChildIndex(this)-1,e>=0?this.parent.children[e]:null):null},e.prototype.getNextSibling=function(){var e;return this.parent?(e=this.parent.getChildIndex(this)+1,e<this.parent.children.length?this.parent.children[e]:null):null},e.prototype.getNodesByProperty=function(e,t){return this.filter(function(n){return n[e]===t})},e.prototype.filter=function(e){var t;return t=[],this.iterate(function(n){return e(n)&&t.push(n),!0}),t},e.prototype.getNextNode=function(e){var t;return null==e&&(e=!0),e&&this.hasChildren()&&this.is_open?this.children[0]:this.parent?(t=this.getNextSibling(),t?t:this.parent.getNextNode(!1)):null},e.prototype.getPreviousNode=function(){var e;return this.parent?(e=this.getPreviousSibling(),e?e.hasChildren()&&e.is_open?e.getLastChild():e:this.getParent()):null},e.prototype.getParent=function(){return this.parent&&this.parent.parent?this.parent:null},e.prototype.getLastChild=function(){var e;return this.hasChildren()?(e=this.children[this.children.length-1],e.hasChildren()&&e.is_open?e.getLastChild():e):null},e}(),t.exports={Node:o,Position:i}},{}],6:[function(e,t,n){var r,o,i,s,a,l,c,u=function(e,t){function n(){this.constructor=e}for(var r in t)d.call(t,r)&&(e[r]=t[r]);return n.prototype=t.prototype,e.prototype=new n,e.__super__=t.prototype,e},d={}.hasOwnProperty;c=e("./node"),l=c.Position,r=jQuery,a=function(){function e(e,t){this.init(e,t)}return e.prototype.init=function(e,t){return this.node=e,this.tree_widget=t,e.element||(e.element=this.tree_widget.element),this.$element=r(e.element)},e.prototype.getUl=function(){return this.$element.children("ul:first")},e.prototype.getSpan=function(){return this.$element.children(".jqtree-element").find("span.jqtree-title")},e.prototype.getLi=function(){return this.$element},e.prototype.addDropHint=function(e){return e===l.INSIDE?new o(this.$element):new s(this.node,this.$element,e)},e.prototype.select=function(){var e,t;return e=this.getLi(),e.addClass("jqtree-selected"),e.attr("aria-selected","true"),t=this.getSpan(),t.attr("tabindex",0)},e.prototype.deselect=function(){var e,t;return e=this.getLi(),e.removeClass("jqtree-selected"),e.attr("aria-selected","false"),t=this.getSpan(),t.attr("tabindex",-1)},e}(),i=function(e){function t(){return t.__super__.constructor.apply(this,arguments)}return u(t,e),t.prototype.open=function(e,t){var n,r;return null==t&&(t=!0),this.node.is_open?void 0:(this.node.is_open=!0,n=this.getButton(),n.removeClass("jqtree-closed"),n.html(""),n.append(this.tree_widget.renderer.opened_icon_element.cloneNode(!1)),r=function(t){return function(){var n,r;return n=t.getLi(),n.removeClass("jqtree-closed"),r=t.getSpan(),r.attr("aria-expanded","true"),e&&e(),t.tree_widget._triggerEvent("tree.open",{node:t.node})}}(this),t?this.getUl().slideDown("fast",r):(this.getUl().show(),r()))},t.prototype.close=function(e){var t,n;return null==e&&(e=!0),this.node.is_open?(this.node.is_open=!1,t=this.getButton(),t.addClass("jqtree-closed"),t.html(""),t.append(this.tree_widget.renderer.closed_icon_element.cloneNode(!1)),n=function(e){return function(){var t,n;return t=e.getLi(),t.addClass("jqtree-closed"),n=e.getSpan(),n.attr("aria-expanded","false"),e.tree_widget._triggerEvent("tree.close",{node:e.node})}}(this),e?this.getUl().slideUp("fast",n):(this.getUl().hide(),n())):void 0},t.prototype.getButton=function(){return this.$element.children(".jqtree-element").find("a.jqtree-toggler")},t.prototype.addDropHint=function(e){return this.node.is_open||e!==l.INSIDE?new s(this.node,this.$element,e):new o(this.$element)},t}(a),o=function(){function e(e){var t,n;t=e.children(".jqtree-element"),n=e.width()-4,this.$hint=r('<span class="jqtree-border"></span>'),t.append(this.$hint),this.$hint.css({width:n,height:t.outerHeight()-4})}return e.prototype.remove=function(){return this.$hint.remove()},e}(),s=function(){function e(e,t,n){this.$element=t,this.node=e,this.$ghost=r('<li class="jqtree_common jqtree-ghost"><span class="jqtree_common jqtree-circle"></span><span class="jqtree_common jqtree-line"></span></li>'),n===l.AFTER?this.moveAfter():n===l.BEFORE?this.moveBefore():n===l.INSIDE&&(e.isFolder()&&e.is_open?this.moveInsideOpenFolder():this.moveInside())}return e.prototype.remove=function(){return this.$ghost.remove()},e.prototype.moveAfter=function(){return this.$element.after(this.$ghost)},e.prototype.moveBefore=function(){return this.$element.before(this.$ghost)},e.prototype.moveInsideOpenFolder=function(){return r(this.node.children[0].element).before(this.$ghost)},e.prototype.moveInside=function(){return this.$element.after(this.$ghost),this.$ghost.addClass("jqtree-inside")},e}(),t.exports={BorderDropHint:o,FolderElement:i,GhostDropHint:s,NodeElement:a}},{"./node":5}],7:[function(e,t,n){var r,o,i,s,a;a=e("./util"),i=a.indexOf,s=a.isInt,r=jQuery,o=function(){function e(e){this.tree_widget=e}return e.prototype.saveState=function(){var e;return e=JSON.stringify(this.getState()),this.tree_widget.options.onSetStateFromStorage?this.tree_widget.options.onSetStateFromStorage(e):this.supportsLocalStorage()?localStorage.setItem(this.getCookieName(),e):r.cookie?(r.cookie.raw=!0,r.cookie(this.getCookieName(),e,{path:"/"})):void 0},e.prototype.getStateFromStorage=function(){var e;return e=this._loadFromStorage(),e?this._parseState(e):null},e.prototype._parseState=function(e){var t;return t=r.parseJSON(e),t&&t.selected_node&&s(t.selected_node)&&(t.selected_node=[t.selected_node]),t},e.prototype._loadFromStorage=function(){return this.tree_widget.options.onGetStateFromStorage?this.tree_widget.options.onGetStateFromStorage():this.supportsLocalStorage()?localStorage.getItem(this.getCookieName()):r.cookie?(r.cookie.raw=!0,r.cookie(this.getCookieName())):null},e.prototype.getState=function(){var e,t;return e=function(e){return function(){var t;return t=[],e.tree_widget.tree.iterate(function(e){return e.is_open&&e.id&&e.hasChildren()&&t.push(e.id),!0}),t}}(this),t=function(e){return function(){var t;return function(){var e,n,r,o;for(r=this.tree_widget.getSelectedNodes(),o=[],e=0,n=r.length;n>e;e++)t=r[e],o.push(t.id);return o}.call(e)}}(this),{open_nodes:e(),selected_node:t()}},e.prototype.setInitialState=function(e){var t;return e?(t=this._openInitialNodes(e.open_nodes),this._selectInitialNodes(e.selected_node),t):!1},e.prototype._openInitialNodes=function(e){var t,n,r,o,i;for(r=!1,t=0,n=e.length;n>t;t++)i=e[t],o=this.tree_widget.getNodeById(i),o&&(o.load_on_demand?r=!0:o.is_open=!0);return r},e.prototype._selectInitialNodes=function(e){var t,n,r,o,i;for(i=0,t=0,n=e.length;n>t;t++)o=e[t],r=this.tree_widget.getNodeById(o),r&&(i+=1,this.tree_widget.select_node_handler.addToSelection(r));return 0!==i},e.prototype.setInitialStateOnDemand=function(e,t){return e?this._setInitialStateOnDemand(e.open_nodes,e.selected_node,t):t()},e.prototype._setInitialStateOnDemand=function(e,t,n){var r,o,i;return o=0,i=function(i){return function(){var s,a,l,c,u;for(l=[],s=0,a=e.length;a>s;s++)u=e[s],c=i.tree_widget.getNodeById(u),c?c.is_loading||(c.load_on_demand?r(c):i.tree_widget._openNode(c,!1)):l.push(u);return e=l,i._selectInitialNodes(t)&&i.tree_widget._refreshElements(),0===o?n():void 0}}(this),r=function(e){return function(t){return o+=1,e.tree_widget._openNode(t,!1,function(){return o-=1,i()})}}(this),i()},e.prototype.getCookieName=function(){return"string"==typeof this.tree_widget.options.saveState?this.tree_widget.options.saveState:"tree"},e.prototype.supportsLocalStorage=function(){var e;return e=function(){var e,t,n;if("undefined"==typeof localStorage||null===localStorage)return!1;try{n="_storage_test",sessionStorage.setItem(n,!0),sessionStorage.removeItem(n)}catch(t){return e=t,!1}return!0},null==this._supportsLocalStorage&&(this._supportsLocalStorage=e()),this._supportsLocalStorage},e.prototype.getNodeIdToBeSelected=function(){var e;return e=this.getStateFromStorage(),e&&e.selected_node?e.selected_node[0]:null},e}(),t.exports=o},{"./util":12}],8:[function(e,t,n){var r,o;r=jQuery,o=function(){function e(e){this.tree_widget=e,this.previous_top=-1,this.is_initialized=!1,this._initScrollParent()}return e.prototype._initScrollParent=function(){var e,t,n;return t=function(e){return function(){var t,n,o,i,s,a;if(t=["overflow","overflow-y"],(o=function(e){var n,o,i,s;for(o=0,i=t.length;i>o;o++)if(n=t[o],"auto"===(s=r.css(e,n))||"scroll"===s)return!0;return!1})(e.tree_widget.$el[0]))return e.tree_widget.$el;for(a=e.tree_widget.$el.parents(),i=0,s=a.length;s>i;i++)if(n=a[i],o(n))return r(n);return null}}(this),n=function(e){return function(){return e.scroll_parent_top=0,e.$scroll_parent=null}}(this),"fixed"===this.tree_widget.$el.css("position")&&n(),e=t(),e&&e.length&&"HTML"!==e[0].tagName?(this.$scroll_parent=e,this.scroll_parent_top=this.$scroll_parent.offset().top):n(),this.is_initialized=!0},e.prototype._ensureInit=function(){return this.is_initialized?void 0:this._initScrollParent()},e.prototype.checkScrolling=function(){var e;return this._ensureInit(),e=this.tree_widget.dnd_handler.hovered_area,e&&e.top!==this.previous_top?(this.previous_top=e.top,this.$scroll_parent?this._handleScrollingWithScrollParent(e):this._handleScrollingWithDocument(e)):void 0},e.prototype._handleScrollingWithScrollParent=function(e){var t;return t=this.scroll_parent_top+this.$scroll_parent[0].offsetHeight-e.bottom,20>t?(this.$scroll_parent[0].scrollTop+=20,this.tree_widget.refreshHitAreas(),this.previous_top=-1):e.top-this.scroll_parent_top<20?(this.$scroll_parent[0].scrollTop-=20,this.tree_widget.refreshHitAreas(),this.previous_top=-1):void 0},e.prototype._handleScrollingWithDocument=function(e){var t;return t=e.top-r(document).scrollTop(),20>t?r(document).scrollTop(r(document).scrollTop()-20):r(window).height()-(e.bottom-r(document).scrollTop())<20?r(document).scrollTop(r(document).scrollTop()+20):void 0},e.prototype.scrollTo=function(e){var t;return this._ensureInit(),this.$scroll_parent?this.$scroll_parent[0].scrollTop=e:(t=this.tree_widget.$el.offset().top,r(document).scrollTop(e+t))},e.prototype.isScrolledIntoView=function(e){var t,n,o,i,s;return this._ensureInit(),t=r(e),this.$scroll_parent?(s=0,i=this.$scroll_parent.height(),o=t.offset().top-this.scroll_parent_top,n=o+t.height()):(s=r(window).scrollTop(),i=s+r(window).height(),o=t.offset().top,n=o+t.height()),i>=n&&o>=s},e}(),t.exports=o},{}],9:[function(e,t,n){var r,o;r=jQuery,o=function(){function e(e){this.tree_widget=e,this.clear()}return e.prototype.getSelectedNode=function(){var e;return e=this.getSelectedNodes(),e.length?e[0]:!1},e.prototype.getSelectedNodes=function(){var e,t,n;if(this.selected_single_node)return[this.selected_single_node];n=[];for(e in this.selected_nodes)t=this.tree_widget.getNodeById(e),t&&n.push(t);return n},e.prototype.getSelectedNodesUnder=function(e){var t,n,r;if(this.selected_single_node)return e.isParentOf(this.selected_single_node)?[this.selected_single_node]:[];r=[];for(t in this.selected_nodes)n=this.tree_widget.getNodeById(t),n&&e.isParentOf(n)&&r.push(n);return r},e.prototype.isNodeSelected=function(e){return e.id?this.selected_nodes[e.id]:this.selected_single_node?this.selected_single_node.element===e.element:!1},e.prototype.clear=function(){return this.selected_nodes={},this.selected_single_node=null},e.prototype.removeFromSelection=function(e,t){if(null==t&&(t=!1),e.id){if(delete this.selected_nodes[e.id],t)return e.iterate(function(t){return function(n){return delete t.selected_nodes[e.id],!0}}(this))}else if(this.selected_single_node&&e.element===this.selected_single_node.element)return this.selected_single_node=null},e.prototype.addToSelection=function(e){return e.id?this.selected_nodes[e.id]=!0:this.selected_single_node=e},e}(),t.exports=o},{}],10:[function(e,t,n){var r,o,i=[].slice;r=jQuery,o=function(){function e(e,t){this.$el=r(e),this.options=r.extend({},this.defaults,t)}return e.prototype.defaults={},e.prototype.destroy=function(){return this._deinit()},e.prototype._init=function(){return null},e.prototype._deinit=function(){return null},e.register=function(t,n){var o,s,a,l,c;return l=function(){return"simple_widget_"+n},c=function(t,n){var o;return o=r.data(t,n),o&&o instanceof e?o:null},s=function(e,n){var o,i,s,a,u,d;for(o=l(),a=0,u=e.length;u>a;a++)i=e[a],s=c(i,o),s||(d=new t(i,n),r.data(i,o)||r.data(i,o,d),d._init());return e},a=function(e){var t,n,o,i,s,a;for(t=l(),s=[],o=0,i=e.length;i>o;o++)n=e[o],a=c(n,t),a&&a.destroy(),s.push(r.removeData(n,t));return s},o=function(t,n,o){var i,s,a,c,u,d;for(c=null,s=0,a=t.length;a>s;s++)i=t[s],u=r.data(i,l()),u&&u instanceof e&&(d=u[n],d&&"function"==typeof d&&(c=d.apply(u,o)));return c},r.fn[n]=function(){var e,n,r,l,c;return r=arguments[0],n=2<=arguments.length?i.call(arguments,1):[],e=this,void 0===r||"object"==typeof r?(c=r,s(e,c)):"string"==typeof r&&"_"!==r[0]?(l=r,"destroy"===l?a(e):"get_widget_class"===l?t:o(e,l,n)):void 0}},e}(),t.exports=o},{}],11:[function(e,t,n){var r,o,i,s,a,l,c,u,d,p,h,f,m,g,_,v,y,b,w,N,x,C,S,k=function(e,t){function n(){this.constructor=e}for(var r in t)E.call(t,r)&&(e[r]=t[r]);return n.prototype=t.prototype,e.prototype=new n,e.__super__=t.prototype,e},E={}.hasOwnProperty;w=e("./version"),x=e("./drag_and_drop_handler"),i=x.DragAndDropHandler,s=x.DragElement,u=x.HitAreasGenerator,a=e("./elements_renderer"),p=e("./key_handler"),h=e("./mouse.widget"),_=e("./save_state_handler"),v=e("./scroll_handler"),y=e("./select_node_handler"),b=e("./simple.widget"),N=e("./node"),f=N.Node,g=N.Position,S=e("./util"),C=e("./node_element"),o=C.BorderDropHint,l=C.FolderElement,c=C.GhostDropHint,m=C.NodeElement,r=jQuery,d=function(e){function t(){return t.__super__.constructor.apply(this,arguments)}return k(t,e),t.prototype.BorderDropHint=o,t.prototype.DragElement=s,t.prototype.DragAndDropHandler=i,t.prototype.ElementsRenderer=a,t.prototype.GhostDropHint=c,t.prototype.HitAreasGenerator=u,t.prototype.Node=f,t.prototype.SaveStateHandler=_,t.prototype.ScrollHandler=v,t.prototype.SelectNodeHandler=y,t.prototype.defaults={autoOpen:!1,saveState:!1,dragAndDrop:!1,selectable:!0,useContextMenu:!0,onCanSelectNode:null,onSetStateFromStorage:null,onGetStateFromStorage:null,onCreateLi:null,onIsMoveHandle:null,onCanMove:null,onCanMoveTo:null,onLoadFailed:null,autoEscape:!0,dataUrl:null,closedIcon:null,openedIcon:"&#x25bc;",slide:!0,nodeClass:f,dataFilter:null,keyboardSupport:!0,openFolderDelay:500,rtl:null,onDragMove:null,onDragStop:null,buttonLeft:!0},t.prototype.toggle=function(e,t){return null==t&&(t=null),null===t&&(t=this.options.slide),e.is_open?this.closeNode(e,t):this.openNode(e,t),this.element},t.prototype.getTree=function(){return this.tree},t.prototype.selectNode=function(e){return this._selectNode(e,!1),this.element},t.prototype._selectNode=function(e,t){var n,r,o,i;if(null==t&&(t=!1),this.select_node_handler){if(n=function(t){return function(){return t.options.onCanSelectNode?t.options.selectable&&t.options.onCanSelectNode(e):t.options.selectable}}(this),o=function(t){return function(){var n;return n=e.parent,n&&n.parent&&!n.is_open?t.openNode(n,!1):void 0}}(this),i=function(e){return function(){return e.options.saveState?e.save_state_handler.saveState():void 0}}(this),!e)return this._deselectCurrentNode(),void i();if(n())return this.select_node_handler.isNodeSelected(e)?t&&(this._deselectCurrentNode(),this._triggerEvent("tree.select",{node:null,previous_node:e})):(r=this.getSelectedNode(),this._deselectCurrentNode(),this.addToSelection(e),this._triggerEvent("tree.select",{node:e,deselected_node:r}),o()),i()}},t.prototype.getSelectedNode=function(){return this.select_node_handler?this.select_node_handler.getSelectedNode():null},t.prototype.toJson=function(){return JSON.stringify(this.tree.getData())},t.prototype.loadData=function(e,t){return this._loadData(e,t),this.element},t.prototype.loadDataFromUrl=function(e,t,n){return"string"===r.type(e)?this._loadDataFromUrl(e,t,n):this._loadDataFromUrl(null,e,t),this.element},t.prototype.reload=function(e){return this._loadDataFromUrl(null,null,e),this.element},t.prototype._loadDataFromUrl=function(e,t,n){var o,i,s,a,l,c;o=null,i=function(e){return function(){return o=t?r(t.element):e.element,o.addClass("jqtree-loading")}}(this),c=function(){return o?o.removeClass("jqtree-loading"):void 0},l=function(){return"string"===r.type(e)&&(e={url:e}),e.method?void 0:e.method="get"},s=function(e){return function(o){return c(),e._loadData(o,t),n&&r.isFunction(n)?n():void 0}}(this),a=function(t){return function(){return l(),r.ajax({url:e.url,data:e.data,type:e.method.toUpperCase(),cache:!1,dataType:"json",success:function(e){var n;return n=r.isArray(e)||"object"==typeof e?e:r.parseJSON(e),
+t.options.dataFilter&&(n=t.options.dataFilter(n)),s(n)},error:function(e){return c(),t.options.onLoadFailed?t.options.onLoadFailed(e):void 0}})}}(this),e||(e=this._getDataUrlInfo(t)),i(),e?r.isArray(e)?s(e):a():c()},t.prototype._loadData=function(e,t){var n,r;return null==t&&(t=null),n=function(e){return function(){var n,r,o,i;if(e.select_node_handler)for(i=e.select_node_handler.getSelectedNodesUnder(t),n=0,r=i.length;r>n;n++)o=i[n],e.select_node_handler.removeFromSelection(o);return null}}(this),r=function(n){return function(){return t.loadFromData(e),t.load_on_demand=!1,t.is_loading=!1,n._refreshElements(t)}}(this),e?(this._triggerEvent("tree.load_data",{tree_data:e}),t?(n(),r()):this._initTree(e),this.isDragging()?this.dnd_handler.refresh():void 0):void 0},t.prototype.getNodeById=function(e){return this.tree.getNodeById(e)},t.prototype.getNodeByName=function(e){return this.tree.getNodeByName(e)},t.prototype.getNodesByProperty=function(e,t){return this.tree.getNodesByProperty(e,t)},t.prototype.openNode=function(e,t){return null==t&&(t=null),null===t&&(t=this.options.slide),this._openNode(e,t),this.element},t.prototype._openNode=function(e,t,n){var r,o;if(null==t&&(t=!0),r=function(e){return function(t,n,r){var o;return o=new l(t,e),o.open(r,n)}}(this),e.isFolder()){if(e.load_on_demand)return this._loadFolderOnDemand(e,t,n);for(o=e.parent;o;)o.parent&&r(o,!1,null),o=o.parent;return r(e,t,n),this._saveState()}},t.prototype._loadFolderOnDemand=function(e,t,n){return null==t&&(t=!0),e.is_loading=!0,this._loadDataFromUrl(null,e,function(r){return function(){return r._openNode(e,t,n)}}(this))},t.prototype.closeNode=function(e,t){return null==t&&(t=null),null===t&&(t=this.options.slide),e.isFolder()&&(new l(e,this).close(t),this._saveState()),this.element},t.prototype.isDragging=function(){return this.dnd_handler?this.dnd_handler.is_dragging:!1},t.prototype.refreshHitAreas=function(){return this.dnd_handler.refresh(),this.element},t.prototype.addNodeAfter=function(e,t){var n;return n=t.addAfter(e),this._refreshElements(t.parent),n},t.prototype.addNodeBefore=function(e,t){var n;return n=t.addBefore(e),this._refreshElements(t.parent),n},t.prototype.addParentNode=function(e,t){var n;return n=t.addParent(e),this._refreshElements(n.parent),n},t.prototype.removeNode=function(e){var t;return t=e.parent,t&&(this.select_node_handler.removeFromSelection(e,!0),e.remove(),this._refreshElements(t)),this.element},t.prototype.appendNode=function(e,t){var n;return t=t||this.tree,n=t.append(e),this._refreshElements(t),n},t.prototype.prependNode=function(e,t){var n;return t||(t=this.tree),n=t.prepend(e),this._refreshElements(t),n},t.prototype.updateNode=function(e,t){var n;return n=t.id&&t.id!==e.id,n&&this.tree.removeNodeFromIndex(e),e.setData(t),n&&this.tree.addNodeToIndex(e),this.renderer.renderFromNode(e),this._selectCurrentNode(),this.element},t.prototype.moveNode=function(e,t,n){var r;return r=g.nameToIndex(n),this.tree.moveNode(e,t,r),this._refreshElements(),this.element},t.prototype.getStateFromStorage=function(){return this.save_state_handler.getStateFromStorage()},t.prototype.addToSelection=function(e){return e&&(this.select_node_handler.addToSelection(e),this._getNodeElementForNode(e).select(),this._saveState()),this.element},t.prototype.getSelectedNodes=function(){return this.select_node_handler.getSelectedNodes()},t.prototype.isNodeSelected=function(e){return this.select_node_handler.isNodeSelected(e)},t.prototype.removeFromSelection=function(e){return this.select_node_handler.removeFromSelection(e),this._getNodeElementForNode(e).deselect(),this._saveState(),this.element},t.prototype.scrollToNode=function(e){var t,n;return t=r(e.element),n=t.offset().top-this.$el.offset().top,this.scroll_handler.scrollTo(n),this.element},t.prototype.getState=function(){return this.save_state_handler.getState()},t.prototype.setState=function(e){return this.save_state_handler.setInitialState(e),this._refreshElements(),this.element},t.prototype.setOption=function(e,t){return this.options[e]=t,this.element},t.prototype.moveDown=function(){return this.key_handler&&this.key_handler.moveDown(),this.element},t.prototype.moveUp=function(){return this.key_handler&&this.key_handler.moveUp(),this.element},t.prototype.getVersion=function(){return w},t.prototype._init=function(){return t.__super__._init.call(this),this.element=this.$el,this.mouse_delay=300,this.is_initialized=!1,this.options.rtl=this._getRtlOption(),this.options.closedIcon||(this.options.closedIcon=this._getDefaultClosedIcon()),this.renderer=new a(this),null!=_?this.save_state_handler=new _(this):this.options.saveState=!1,null!=y&&(this.select_node_handler=new y(this)),null!=i?this.dnd_handler=new i(this):this.options.dragAndDrop=!1,null!=v&&(this.scroll_handler=new v(this)),null!=p&&null!=y&&(this.key_handler=new p(this)),this._initData(),this.element.click(r.proxy(this._click,this)),this.element.dblclick(r.proxy(this._dblclick,this)),this.options.useContextMenu?this.element.bind("contextmenu",r.proxy(this._contextmenu,this)):void 0},t.prototype._deinit=function(){return this.element.empty(),this.element.unbind(),this.key_handler&&this.key_handler.deinit(),this.tree=null,t.__super__._deinit.call(this)},t.prototype._initData=function(){return this.options.data?this._loadData(this.options.data):this._loadDataFromUrl(this._getDataUrlInfo())},t.prototype._getDataUrlInfo=function(e){var t,n;return t=this.options.dataUrl||this.element.data("url"),n=function(n){return function(){var r,o,i;return i={url:t},e&&e.id?(r={node:e.id},i.data=r):(o=n._getNodeIdToBeSelected(),o&&(r={selected_node:o},i.data=r)),i}}(this),r.isFunction(t)?t(e):"string"===r.type(t)?n():t},t.prototype._getNodeIdToBeSelected=function(){return this.options.saveState?this.save_state_handler.getNodeIdToBeSelected():null},t.prototype._initTree=function(e){var t,n;return t=function(e){return function(){return e.is_initialized?void 0:(e.is_initialized=!0,e._triggerEvent("tree.init"))}}(this),this.tree=new this.options.nodeClass(null,!0,this.options.nodeClass),this.select_node_handler&&this.select_node_handler.clear(),this.tree.loadFromData(e),n=this._setInitialState(),this._refreshElements(),n?this._setInitialStateOnDemand(t):t()},t.prototype._setInitialState=function(){var e,t,n,r,o;return o=function(e){return function(){var t,n;return e.options.saveState&&e.save_state_handler?(n=e.save_state_handler.getStateFromStorage(),n?(t=e.save_state_handler.setInitialState(n),[!0,t]):[!1,!1]):[!1,!1]}}(this),e=function(e){return function(){var t,n;return e.options.autoOpen===!1?!1:(t=e._getAutoOpenMaxLevel(),n=!1,e.tree.iterate(function(e,r){return e.load_on_demand?(n=!0,!1):e.hasChildren()?(e.is_open=!0,r!==t):!1}),n)}}(this),r=o(),t=r[0],n=r[1],t||(n=e()),n},t.prototype._setInitialStateOnDemand=function(e){var t,n;return n=function(t){return function(){var n;return t.options.saveState&&t.save_state_handler?(n=t.save_state_handler.getStateFromStorage(),n?(t.save_state_handler.setInitialStateOnDemand(n,e),!0):!1):!1}}(this),t=function(t){return function(){var n,r,o,i;return o=t._getAutoOpenMaxLevel(),r=0,n=function(e){return r+=1,t._openNode(e,!1,function(){return r-=1,i()})},(i=function(){return t.tree.iterate(function(e,r){return e.load_on_demand?(e.is_loading||n(e),!1):(t._openNode(e,!1),r!==o)}),0===r?e():void 0})()}}(this),n()?void 0:t()},t.prototype._getAutoOpenMaxLevel=function(){return this.options.autoOpen===!0?-1:parseInt(this.options.autoOpen)},t.prototype._refreshElements=function(e){return null==e&&(e=null),this.renderer.render(e),this._triggerEvent("tree.refresh")},t.prototype._click=function(e){var t,n,r;if(t=this._getClickTarget(e.target)){if("button"===t.type)return this.toggle(t.node,this.options.slide),e.preventDefault(),e.stopPropagation();if("label"===t.type&&(r=t.node,n=this._triggerEvent("tree.click",{node:r,click_event:e}),!n.isDefaultPrevented()))return this._selectNode(r,!0)}},t.prototype._dblclick=function(e){var t;return t=this._getClickTarget(e.target),t&&"label"===t.type?this._triggerEvent("tree.dblclick",{node:t.node,click_event:e}):void 0},t.prototype._getClickTarget=function(e){var t,n,o,i;if(o=r(e),t=o.closest(".jqtree-toggler"),t.length){if(i=this._getNode(t))return{type:"button",node:i}}else if(n=o.closest(".jqtree-element"),n.length&&(i=this._getNode(n)))return{type:"label",node:i};return null},t.prototype._getNode=function(e){var t;return t=e.closest("li.jqtree_common"),0===t.length?null:t.data("node")},t.prototype._getNodeElementForNode=function(e){return e.isFolder()?new l(e,this):new m(e,this)},t.prototype._getNodeElement=function(e){var t;return t=this._getNode(e),t?this._getNodeElementForNode(t):null},t.prototype._contextmenu=function(e){var t,n;return t=r(e.target).closest("ul.jqtree-tree .jqtree-element"),t.length&&(n=this._getNode(t))?(e.preventDefault(),e.stopPropagation(),this._triggerEvent("tree.contextmenu",{node:n,click_event:e}),!1):void 0},t.prototype._saveState=function(){return this.options.saveState?this.save_state_handler.saveState():void 0},t.prototype._mouseCapture=function(e){return this.options.dragAndDrop?this.dnd_handler.mouseCapture(e):!1},t.prototype._mouseStart=function(e){return this.options.dragAndDrop?this.dnd_handler.mouseStart(e):!1},t.prototype._mouseDrag=function(e){var t;return this.options.dragAndDrop?(t=this.dnd_handler.mouseDrag(e),this.scroll_handler&&this.scroll_handler.checkScrolling(),t):!1},t.prototype._mouseStop=function(e){return this.options.dragAndDrop?this.dnd_handler.mouseStop(e):!1},t.prototype._triggerEvent=function(e,t){var n;return n=r.Event(e),r.extend(n,t),this.element.trigger(n),n},t.prototype.testGenerateHitAreas=function(e){return this.dnd_handler.current_item=this._getNodeElementForNode(e),this.dnd_handler.generateHitAreas(),this.dnd_handler.hit_areas},t.prototype._selectCurrentNode=function(){var e,t;return e=this.getSelectedNode(),e&&(t=this._getNodeElementForNode(e))?t.select():void 0},t.prototype._deselectCurrentNode=function(){var e;return e=this.getSelectedNode(),e?this.removeFromSelection(e):void 0},t.prototype._getDefaultClosedIcon=function(){return this.options.rtl?"&#x25c0;":"&#x25ba;"},t.prototype._getRtlOption=function(){var e;return null!==this.options.rtl?this.options.rtl:(e=this.element.data("rtl"),null!=e&&e!==!1)},t}(h),d.getModule=function(e){var t;return t={node:N,util:S},t[e]},b.register(d,"tree")},{"./drag_and_drop_handler":1,"./elements_renderer":2,"./key_handler":3,"./mouse.widget":4,"./node":5,"./node_element":6,"./save_state_handler":7,"./scroll_handler":8,"./select_node_handler":9,"./simple.widget":10,"./util":12,"./version":13}],12:[function(e,t,n){var r,o,i,s,a;r=function(e,t){var n,r,o,i;for(n=r=0,o=e.length;o>r;n=++r)if(i=e[n],i===t)return n;return-1},s=function(e,t){return e.indexOf?e.indexOf(t):r(e,t)},a=function(e){return"number"==typeof e&&e%1===0},i=function(e){return(""+e).replace(/&/g,"&amp;").replace(/</g,"&lt;").replace(/>/g,"&gt;").replace(/"/g,"&quot;").replace(/'/g,"&#x27;").replace(/\//g,"&#x2F;")},o=function(e){return e?"true":"false"},t.exports={_indexOf:r,getBoolString:o,html_escape:i,indexOf:s,isInt:a}},{}],13:[function(e,t,n){t.exports="1.3.0"},{}]},{},[11]);
